% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_rm.R
\name{add_ancestors}
\alias{add_ancestors}
\title{Add ancestor records for an individual}
\usage{
add_ancestors(x, xref, num_gen, inc_sex = TRUE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record to add ancestors for.}

\item{num_gen}{The number of generations to create ancestors for.}

\item{inc_sex}{Whether to populate the sex of the ancestors. This will ensure
that there is one male and one female parent. Otherwise the sex will be
assigned as "U" (undetermined).}
}
\value{
A gedcom object with additional ancestor records.
}
\description{
This function adds placeholder Individual records for ancestors going back a specific
number of generations.
}
\details{
This function may also create Family records and will
not modify existing ancestors.
}
