% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_children}
\alias{add_children}
\title{Create multiple children for a Family Group}
\usage{
add_children(x, xref, sexes)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of a Family Group record.}

\item{sexes}{A character string giving the sexes of each child. For example,
"FFM" to add two daughters and one son.}
}
\value{
A gedcom object with additional child records.
}
\description{
Create multiple children for a Family Group
}
