% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_spouse}
\alias{add_spouse}
\title{Add a spouse for an individual}
\usage{
add_spouse(x, xref, sex = "U")
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{sex}{The sex of the spouse.}
}
\value{
A gedcom object with additional spouse and Family Group records.
}
\description{
This creates a record for a spouse and their Family Group record.
}
