% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_address.R
\name{Address}
\alias{Address}
\title{Create an address object}
\usage{
Address(
  full = character(0),
  city = character(0),
  state = character(0),
  postal_code = character(0),
  country = character(0),
  adr1 = character(0),
  adr2 = character(0),
  adr3 = character(0)
)
}
\arguments{
\item{full}{A full address as it would appear on a mailing label, with lines separated
by semi-colon and a space. For example:
"The White House; 1600 Pennsylvania Avenue N.W.; Washington, DC 20500; United States of America"}

\item{city}{The city component of the address.}

\item{state}{The state component of the address.}

\item{postal_code}{The postal code component of the address.}

\item{country}{The country component of the address.}

\item{adr1}{Deprecated.}

\item{adr2}{Deprecated.}

\item{adr3}{Deprecated.}
}
\value{
An S7 object representing a GEDCOM ADDRESS_STRUCTURE.
}
\description{
Create an address object
}
