% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_association.R
\name{Association}
\alias{Association}
\title{Create an association object}
\usage{
Association(
  indi_xref = "@VOID@",
  indi_phrase = character(0),
  relation_is = character(0),
  relation_phrase = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list()
)
}
\arguments{
\item{indi_xref}{The cross-reference identifier of an individual record. If the individual
does not have a record, then this can be left blank and a void xref will be used. However,
you should define an @indi_phrase.}

\item{indi_phrase}{Textual information that cannot be expressed in the @indi_xref.}

\item{relation_is}{The nature of the association. This must be a value from \code{val_roles()}.
If a value of "OTHER" is used, a @relation_phrase must be given.}

\item{relation_phrase}{Textual information that cannot be expressed in the relation.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records.}
}
\value{
An S7 object representing a GEDCOM ASSOCIATION_STRUCTURE.
}
\description{
Create an association object
}
