% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{check_for_xref_mentions}
\alias{check_for_xref_mentions}
\title{Check gedcom lines for inappropriate mentions of xrefs}
\usage{
check_for_xref_mentions(lines)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}
}
\value{
Nothing. If applicable, the function will result in a warning if
inappropriate mentions of xrefs are found.
}
\description{
Check gedcom lines for inappropriate mentions of xrefs
}
\keyword{internal}
