% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_validators.R
\name{chk_input_pattern}
\alias{chk_input_pattern}
\title{Validate the regex pattern(s) of a character vector}
\usage{
chk_input_pattern(input, pattern)
}
\arguments{
\item{input}{The input.}

\item{pattern}{The regex pattern each element must match.}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate the regex pattern(s) of a character vector
}
\keyword{internal}
