% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_validators.R
\name{chk_input_phrase}
\alias{chk_input_phrase}
\title{Validate the existence of a subordinate phrase}
\usage{
chk_input_phrase(input, name, parent, parent_name, parent_val)
}
\arguments{
\item{input}{The value of the phrase input.}

\item{name}{The phrase input name.}

\item{parent}{The parent input.}

\item{parent_name}{The name of the parent input used in any error messages.}

\item{parent_val}{The value of the parent input which requires a subordinate phrase.}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate the existence of a subordinate phrase
}
\keyword{internal}
