% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_validators.R
\name{chk_input_S7classes}
\alias{chk_input_S7classes}
\title{Validate a list to ensure it contains elements of the correct S7 class}
\usage{
chk_input_S7classes(input_lst, target_class)
}
\arguments{
\item{input_lst}{The list of S7 objects to check.}

\item{target_class}{The S7 class that the elements of the list should contain.}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate a list to ensure it contains elements of the correct S7 class
}
\keyword{internal}
