% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_validators.R
\name{chk_input_size}
\alias{chk_input_size}
\title{Validate the size of an input}
\usage{
chk_input_size(
  input,
  name,
  min_len = NULL,
  max_len = NULL,
  min_val = NULL,
  max_val = NULL
)
}
\arguments{
\item{input}{The input.}

\item{name}{The name of the input used in any error messages.}

\item{min_len}{The minimum number of elements the input should have.}

\item{max_len}{The maximum number of elements the input should have.}

\item{min_val}{The minimum number of characters or value the input should have
(depending on input type).}

\item{max_val}{The maximum number of characters or value the input should have
(depending on input type).}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate the size of an input
}
