% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_validators.R
\name{chk_whole_number}
\alias{chk_whole_number}
\title{Validate an numeric input to be a whole number}
\usage{
chk_whole_number(input, name)
}
\arguments{
\item{input}{The input.}

\item{name}{The name of the input used in any error messages.}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate an numeric input to be a whole number
}
