% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{chronify}
\alias{chronify}
\title{Force a vector to be a length 1 character vector}
\usage{
chronify(x)
}
\arguments{
\item{x}{An atomic vector of any length.}
}
\value{
A character vector of length one. It is either an empty string for a
zero length input, or takes the value of the first element.
}
\description{
The name comes from 'Chr-one-ify'.
}
