% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_address.R
\name{class_address}
\alias{class_address}
\title{Create an address object}
\usage{
class_address(
  full = class_missing,
  city = class_missing,
  state = class_missing,
  postal_code = class_missing,
  country = class_missing
)
}
\arguments{
\item{full}{A full address as it would appear on a mailing label, with lines separated
by \n. For example:
"The White House\n1600 Pennsylvania Avenue, N.W.\nWashington, DC 20500\nUnited States of America"}

\item{city}{Optional. The city component of the address.}

\item{state}{Optional. The state component of the address.}

\item{postal_code}{Optional. The postal code component of the address.}

\item{country}{Optional. The country component of the address.}
}
\value{
An S7 object representing a GEDCOM ADDRESS_STRUCTURE.
}
\description{
Create an address object
}
