% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_family_link.R
\name{class_child_family_link}
\alias{class_child_family_link}
\title{Create a family link (as child) object}
\usage{
class_child_family_link(
  fam_xref = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  pedigree = class_missing,
  pedigree_phrase = class_missing,
  confidence = class_missing,
  confidence_phrase = class_missing
)
}
\arguments{
\item{fam_xref}{The cross-reference identifier of a family record.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{pedigree}{TODO}

\item{pedigree_phrase}{TODO}

\item{confidence}{TODO}

\item{confidence_phrase}{TODO}
}
\value{
An S7 object representing a GEDCOM family link as a child.
}
\description{
Create a family link (as child) object
}
