% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_citation.R
\name{class_citation}
\alias{class_citation}
\title{Create a source citation object}
\usage{
class_citation(
  sour_uid = class_missing,
  where = class_missing,
  date = class_missing,
  source_text = class_missing,
  event_type = class_missing,
  event_phrase = class_missing,
  role = class_missing,
  role_phrase = class_missing,
  certainty = class_missing,
  media_links = class_missing,
  note_uids = class_missing,
  notes = class_missing
)
}
\arguments{
\item{sour_uid}{The UID of a source record.}

\item{notes}{}
}
\value{
An S7 object representing a GEDCOM SOURCE_CITATION.
}
\description{
Create a source citation object
}
