% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_date.R
\name{class_date_approx}
\alias{class_date_approx}
\title{Create a GEDCOM Approximate Date object}
\usage{
class_date_approx(
  date = class_missing,
  about = class_missing,
  calc = class_missing,
  est = class_missing
)
}
\arguments{
\item{date}{A Gregorian date given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{about}{Whether the date is near to the date given.}

\item{calc}{Whether the date is calculated from other values.}

\item{est}{Whether the date is near to the date given, and is calculated from other values.}
}
\value{
An S7 object representing a GEDCOM Approximate Date.
}
\description{
Create a GEDCOM Approximate Date object
}
