% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_date.R
\name{class_date_period}
\alias{class_date_period}
\title{Create a GEDCOM Date Period object}
\usage{
class_date_period(start_date = class_missing, end_date = class_missing)
}
\arguments{
\item{start_date}{The start of the period given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{end_date}{The end of the period given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}
}
\value{
An S7 object representing a GEDCOM Date Period.
}
\description{
Create a GEDCOM Date Period object
}
