% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_date.R
\name{class_date_range}
\alias{class_date_range}
\title{Create a GEDCOM Date Range object}
\usage{
class_date_range(start_date = character(0), end_date = character(0))
}
\arguments{
\item{start_date}{The start of the period/range given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{end_date}{The end of the period/range given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}
}
\value{
An S7 object representing a GEDCOM Date Range.
}
\description{
Create a GEDCOM Date Range object
}
