% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_date.R
\name{class_date_value}
\alias{class_date_value}
\title{Create a GEDCOM Date Value object}
\usage{
class_date_value(
  date = class_missing,
  date_phrase = class_missing,
  time = class_missing
)
}
\arguments{
\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_value} object.}

\item{date_phrase}{Textual information that cannot be expressed in the date.}

\item{time}{The time given either as a formatted GEDCOM string, or a
\code{class_time} object.}
}
\value{
An S7 object representing a GEDCOM Date Value.
}
\description{
Create a GEDCOM Date Value object
}
