% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_fact_fam.R
\name{class_event_fam}
\alias{class_event_fam}
\title{Create a family event object}
\usage{
class_event_fam(
  fact_type = character(0),
  fact_val = character(0),
  fact_desc = character(0),
  date = character(0),
  place = character(0),
  address = character(0),
  phone_numbers = character(0),
  emails = character(0),
  faxes = character(0),
  web_pages = character(0),
  agency = character(0),
  relig_affil = character(0),
  cause = character(0),
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  date_sort = character(0),
  associations = list(),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  unique_ids = character(0),
  husb_age = character(0),
  husb_age_phrase = character(0),
  wife_age = character(0),
  wife_age_phrase = character(0)
)
}
\arguments{
\item{fact_type}{TODO}

\item{fact_val}{TODO}

\item{fact_desc}{TODO}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_value} object.}

\item{place}{see territory}

\item{address}{The address given either as a \code{class_address} object or as a character string.
This would be as written on a mailing label with new lines separated by \\n.}

\item{phone_numbers}{A character vector of phone numbers.}

\item{emails}{A character vector of email addresses.}

\item{faxes}{A character vector of fax numbers.}

\item{web_pages}{A character vector of web page URLs.}

\item{agency}{The organization, institution, corporation, person, or other entity that
has responsibility for the associated fact. Examples are an employer of a person of an
associated occupation, or an educational establishment that has awarded a scholastic award.}

\item{relig_affil}{A religious denomination associated with the fact.}

\item{cause}{Used in special cases to record the reasons which precipitated the fact (e.g. cause of death).}

\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{date_sort}{The date given either as a formatted GEDCOM string, or a
\code{class_date_sort} object.}

\item{associations}{Associated individuals. This can either be a \code{class_association} object or a list of them.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Associated multimedia. This can either be a \code{class_media_link} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}
}
\value{
An S7 object representing a GEDCOM FAMILY_EVENT_STRUCTURE.
}
\description{
Create a family event object
}
