% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_fact_indi.R
\name{class_event_indi}
\alias{class_event_indi}
\title{Create an individual event object}
\usage{
class_event_indi(
  fact_type = class_missing,
  fact_val = class_missing,
  fact_desc = class_missing,
  date = class_missing,
  place = class_missing,
  address = class_missing,
  phone_numbers = class_missing,
  emails = class_missing,
  faxes = class_missing,
  web_pages = class_missing,
  agency = class_missing,
  relig_affil = class_missing,
  cause = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  date_sort = class_missing,
  associations = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  media_links = class_missing,
  unique_ids = class_missing,
  age = class_missing,
  age_phrase = class_missing,
  fam_xref = class_missing,
  adop_parent = class_missing,
  adop_parent_phrase = class_missing
)
}
\arguments{
\item{fact_type}{}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_value} object.}

\item{place}{see territory}

\item{address}{The address given either as a \code{class_address} object or as a character string.
This would be as written on a mailing label with new lines separated by \n.}

\item{phone_numbers}{A character vector of phone numbers.}

\item{emails}{A character vector of email addresses.}

\item{faxes}{A character vector of fax numbers.}

\item{web_pages}{A character vector of web page URLs.}

\item{agency}{The organization, institution, corporation, person, or other entity that
has responsibility for the associated fact. Examples are an employer of a person of an
associated occupation, or an educational establishment that has awarded a scholastic award.}

\item{cause}{Used in special cases to record the reasons which precipitated the fact (e.g. cause of death).}

\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{date_sort}{The date given either as a formatted GEDCOM string, or a
\code{class_date_sort} object.}

\item{associations}{Associated individuals. This can either be a \code{class_association} object or a list of them.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Associated multimedia. This can either be a \code{class_media_link} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{fam_xref}{The cross-reference identifier of a family record.}
}
\value{
An S7 object representing a GEDCOM INDIVIDUAL_EVENT_STRUCTURE.
}
\description{
Create an individual event object
}
