% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{class_gedcom_header}
\alias{class_gedcom_header}
\title{Create a GEDCOM header object}
\usage{
class_gedcom_header(
  gedcom_version = class_missing,
  ext_tags = class_missing,
  source = class_missing,
  destination = class_missing,
  creation_date = class_missing,
  creation_time = class_missing,
  subm_xref = class_missing,
  gedcom_copyright = class_missing,
  default_language = class_missing,
  default_place_form = class_missing,
  notes = class_missing,
  note_xrefs = class_missing
)
}
\arguments{
\item{creation_date}{The creation date of the file given either as a formatted GEDCOM string, or a
\code{class_date_exact} object.}

\item{creation_time}{The creation time of the file given either as a formatted GEDCOM string, or a
\code{class_time} object.}

\item{subm_xref}{The cross-reference identifier of a submitter record.}

\item{default_place_form}{}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}
}
\value{
An S7 object representing a GEDCOM HEADER.
}
\description{
Create a GEDCOM header object
}
