% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{class_gedcom_source}
\alias{class_gedcom_source}
\title{Create a GEDCOM source object}
\usage{
class_gedcom_source(
  product_id = character(0),
  product_name = character(0),
  product_version = character(0),
  business_name = character(0),
  business_address = character(0),
  phone_numbers = character(0),
  emails = character(0),
  faxes = character(0),
  web_pages = character(0),
  data_name = character(0),
  data_pubdate = character(0),
  data_pubtime = character(0),
  data_copyright = character(0)
)
}
\arguments{
\item{product_id}{TODO}

\item{product_name}{TODO}

\item{product_version}{TODO}

\item{business_name}{TODO}

\item{business_address}{TODO}

\item{phone_numbers}{A character vector of phone numbers.}

\item{emails}{A character vector of email addresses.}

\item{faxes}{A character vector of fax numbers.}

\item{web_pages}{A character vector of web page URLs.}

\item{data_name}{TODO}

\item{data_pubdate}{TODO}

\item{data_pubtime}{TODO}

\item{data_copyright}{TODO}
}
\value{
An S7 object representing a GEDCOM HEAD.SOUR.
}
\description{
Create a GEDCOM source object
}
