% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{class_gedcomS7}
\alias{class_gedcomS7}
\title{Create a GEDCOM object}
\usage{
class_gedcomS7(
  gedcom_version = class_missing,
  ext_tags = class_missing,
  source = class_missing,
  destination = class_missing,
  creation_date = class_missing,
  creation_time = class_missing,
  subm_xref = class_missing,
  gedcom_copyright = class_missing,
  default_language = class_missing,
  default_place_form = class_missing,
  notes = class_missing,
  note_xrefs = class_missing,
  update_change_dates = class_missing,
  add_creation_dates = class_missing,
  subm = class_missing,
  indi = class_missing,
  fam = class_missing,
  sour = class_missing,
  repo = class_missing,
  media = class_missing,
  note = class_missing
)
}
\arguments{
\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}
}
\value{
An S7 object representing a GEDCOM file.
}
\description{
Create a GEDCOM object
}
