% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_media_file.R
\name{class_media_file}
\alias{class_media_file}
\title{Create a media file object}
\usage{
class_media_file(
  location = class_missing,
  title = class_missing,
  media_type = class_missing,
  medium = class_missing,
  medium_phrase = class_missing,
  media_alt = class_missing
)
}
\arguments{
\item{title}{The title of the multimedia record. This will supersede any title given
in the record.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{media_alt}{A named vector of the media in alternative media forms, c(form = location)}
}
\value{
An S7 object representing a GEDCOM multimedia file substructure.
}
\description{
Create a media file object
}
