% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_media_link.R
\name{class_media_link}
\alias{class_media_link}
\title{Create a multimedia link object}
\usage{
class_media_link(
  media_uid = class_missing,
  title = class_missing,
  crop = class_missing,
  top = class_missing,
  left = class_missing,
  height = class_missing,
  width = class_missing
)
}
\arguments{
\item{media_uid}{The UID of a multimedia record.}

\item{title}{The title of the multimedia record. This will supercede any title given
in the record.}

\item{crop}{Whether to crop the multimedia to a specific area.}

\item{top}{The number of pixels to omit from the top side of the image.}

\item{left}{The number of pixels to omit from the left side of the image.}

\item{height}{The height in pixels of the cropped region.}

\item{width}{The width in pixels of the cropped region.}
}
\value{
An S7 object representing a GEDCOM MULTIMEDIA_LINK.
}
\description{
Create a multimedia link object
}
\details{
The properties @left and @top indicate the top left corner of the region to display.
The properties @width and @height indicate the dimensions of the region to display.

If the multimedia record contains multiple files, then the crop parameters only
applies to the first file.
}
