% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_note.R
\name{class_note}
\alias{class_note}
\title{Create a note structure object}
\usage{
class_note(
  text = class_missing,
  language = class_missing,
  media_type = class_missing,
  alt_text = class_missing
)
}
\arguments{
\item{text}{A character string. New lines are created with \n.}

\item{language}{Optional. Language tags as defined in BCP 47.}

\item{media_type}{Optional. The media type as defined in RFC 2045.}

\item{alt_text}{Optional. A \code{class_translation_txt} object or a list of them.
One for each alternate translation of the text.}
}
\value{
An S7 object representing a GEDCOM NOTE_STRUCTURE.
}
\description{
Create a note structure object
}
\details{
The shared note (SNOTE) alternative of this structure is defined
separately in relevant structures.

In addition, this class does not include source citations as it
results in infinite nesting.
https://github.com/RConsortium/OOP-WG/issues/250
}
