% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_personal_name.R
\name{class_personal_name}
\alias{class_personal_name}
\title{Create a personal name object}
\usage{
class_personal_name(
  pers_name = class_missing,
  name_type = class_missing,
  type_phrase = class_missing,
  name_pieces = class_missing,
  name_translations = class_missing,
  notes = class_missing,
  note_xrefs = class_missing,
  citations = class_missing
)
}
\arguments{
\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}
}
\value{
An S7 object representing a GEDCOM PERSONAL_NAME_STRUCTURE.
}
\description{
Create a personal name object
}
