% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_personal_name.R
\name{class_personal_name_tran}
\alias{class_personal_name_tran}
\title{Create a name translation object}
\usage{
class_personal_name_tran(
  pers_name = class_missing,
  language = class_missing,
  name_pieces = class_missing
)
}
\arguments{
\item{pers_name}{TODO}

\item{language}{A character string of language tags as defined in BCP 47.}

\item{name_pieces}{A \code{class_name_pieces} object defining the pieces of the full name.}
}
\value{
An S7 object representing a GEDCOM personal name translation substructure.
}
\description{
Create a name translation object
}
