% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_place.R
\name{class_place}
\alias{class_place}
\title{Create a place structure object}
\usage{
class_place(
  place_name = class_missing,
  place_form = class_missing,
  language = class_missing,
  place_translations = class_missing,
  lat_long = class_missing,
  ext_ids = class_missing,
  note_xrefs = class_missing,
  notes = class_missing
)
}
\arguments{
\item{language}{Language tags as defined in BCP 47.}

\item{ext_ids}{dsfse}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}
}
\value{
An S7 object representing a GEDCOM PLACE_STRUCTURE.
}
\description{
Create a place structure object
}
