% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record.R
\name{class_record}
\alias{class_record}
\title{Create a base record object}
\usage{
class_record(
  xref = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  user_ids = class_missing,
  unique_ids = class_missing,
  ext_ids = class_missing,
  created = class_missing,
  updated = class_missing
)
}
\arguments{
\item{unique_ids}{}
}
\value{
An S7 object containing common elements of a GEDCOM record.
}
\description{
Create a base record object
}
