% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_note.R
\name{class_record_note}
\alias{class_record_note}
\title{Create a shared note record object}
\usage{
class_record_note(
  xref = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  user_ids = class_missing,
  unique_ids = class_missing,
  ext_ids = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  media_links = class_missing,
  created = class_missing,
  updated = class_missing,
  text = class_missing,
  media_type = class_missing,
  language = class_missing,
  translations = class_missing
)
}
\arguments{
\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{user_ids}{fwe}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{dsfse}

\item{note_xrefs}{Not used.}

\item{notes}{Not used.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Not used.}

\item{updated}{}

\item{text}{A character string. New lines are created with \n.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{language}{Language tags as defined in BCP 47.}

\item{translations}{A \code{class_translation_txt} object or a list of them.
One for each alternate translation of the text.}
}
\value{
An S7 object representing a GEDCOM SHARED_NOTE_RECORD.
}
\description{
Create a shared note record object
}
