% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_repo.R
\name{class_record_repo}
\alias{class_record_repo}
\title{Create a repository record object}
\usage{
class_record_repo(
  xref = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  user_ids = class_missing,
  unique_ids = class_missing,
  ext_ids = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  media_links = class_missing,
  created = class_missing,
  updated = class_missing,
  repo_name = class_missing,
  address = class_missing,
  phone_numbers = class_missing,
  emails = class_missing,
  faxes = class_missing,
  web_pages = class_missing
)
}
\arguments{
\item{xref}{}

\item{user_ids}{fwe}

\item{unique_ids}{ds}

\item{ext_ids}{dsfse}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Not used.}

\item{media_links}{Not used.}

\item{updated}{}

\item{phone_numbers}{}
}
\value{
An S7 object representing a GEDCOM REPOSITORY_RECORD.
}
\description{
Create a repository record object
}
