% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_subm.R
\name{class_record_subm}
\alias{class_record_subm}
\title{Create a submitter record object}
\usage{
class_record_subm(
  xref = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  user_ids = class_missing,
  unique_ids = class_missing,
  ext_ids = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  media_links = class_missing,
  created = class_missing,
  updated = class_missing,
  subm_name = class_missing,
  address = class_missing,
  phone_numbers = class_missing,
  emails = class_missing,
  faxes = class_missing,
  web_pages = class_missing,
  languages = class_missing
)
}
\arguments{
\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{user_ids}{A character vector of user-generated identifiers. The type of the identifiers can
be given in the vector names, e.g. c("Driving license number" = "ABC123")}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{A named character vector of identifiers maintained by an external authority.
The names must be given as a URI. See the GEDCOM specification for more information.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Not used.}

\item{media_links}{Associated multimedia. This can either be a \code{class_media_link} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{created}{A \code{class_creation_date} object containing the date the record was created.
Creating an object with no parameters sets the date to today.}

\item{updated}{A \code{class_change_date} object containing the date the record was updated.
Creating an object with no parameters sets the date to today.}

\item{subm_name}{The name of the submitter.}

\item{address}{The address given either as a \code{class_address} object or as a character string.
This would be as written on a mailing label with new lines separated by \n.}

\item{phone_numbers}{}

\item{languages}{A character vector of language tags as defined in BCP 47.}
}
\value{
An S7 object representing a GEDCOM SUBMITTER_RECORD.
}
\description{
Create a submitter record object
}
