% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{class_repository_citation}
\alias{class_repository_citation}
\title{Create a repository citation object}
\usage{
class_repository_citation(
  repo_xref = class_missing,
  notes = class_missing,
  note_xrefs = class_missing,
  call_numbers = class_missing
)
}
\arguments{
\item{repo_xref}{The cross-reference identifier of a repository record. If the repository
does not have a record, then this can be left blank and the value "@VOID@" will be used. However,
you should describe the repository in @notes.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{call_numbers}{TODO}
}
\value{
An S7 object representing a GEDCOM SOURCE_REPOSITORY_CITATION.
}
\description{
Create a repository citation object
}
