% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{class_source_call_number}
\alias{class_source_call_number}
\title{Create a source call number object}
\usage{
class_source_call_number(
  call_number = character(0),
  medium = character(0),
  medium_phrase = character(0)
)
}
\arguments{
\item{call_number}{The call number.}

\item{medium}{TODO}

\item{medium_phrase}{TODO}
}
\value{
An S7 object representing the CALN substructure of a GEDCOM
SOURCE_REPOSITORY_CITATION.
}
\description{
Create a source call number object
}
