% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ordinance.R
\name{class_spouse_sealing}
\alias{class_spouse_sealing}
\title{Create a spouse sealing object}
\usage{
class_spouse_sealing(
  date = character(0),
  temple_name = character(0),
  place = character(0),
  ord_state = character(0),
  state_date = character(0),
  state_time = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list()
)
}
\arguments{
\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_value} object.}

\item{temple_name}{TODO}

\item{place}{see territory}

\item{ord_state}{TODO}

\item{state_date}{The ordinance date given either as a formatted GEDCOM string, or a
\code{class_date_exact} object.}

\item{state_time}{The ordinance time given either as a formatted GEDCOM string, or a
\code{class_time} object.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}
}
\value{
An S7 object representing a GEDCOM LDS_SPOUSE_SEALING.
}
\description{
Create a spouse sealing object
}
