% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_time.R
\name{class_time}
\alias{class_time}
\title{Create a time object}
\usage{
class_time(
  hour = class_missing,
  minute = class_missing,
  second = class_missing,
  fraction = class_missing,
  utc = class_missing
)
}
\arguments{
\item{hour}{The hour of the day given as an integer between 0 and 23.}

\item{minute}{The minute of the hour given as an integer between 0 and 59.}

\item{second}{The second of the minute given as an integer between 0 and 59.}

\item{fraction}{The fraction of the second given as an integer.}

\item{utc}{Whether the time is in Coordinated Universal Time (UTC) (TRUE, the default) or
is in local time (FALSE).}
}
\value{
An S7 object representing a GEDCOM time.
}
\description{
Create a time object
}
