% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_import.R
\name{combine_cont_lines}
\alias{combine_cont_lines}
\title{Combine continuation lines into the parent line}
\usage{
combine_cont_lines(lines)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}
}
\value{
A new character vector of gedcom lines.
}
\description{
Combine continuation lines into the parent line
}
\keyword{internal}
