% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_import.R
\name{combine_gedcom_values}
\alias{combine_gedcom_values}
\title{Convert the GEDCOM grammar to the GEDCOM form}
\usage{
combine_gedcom_values(lines)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}
}
\value{
A new character vector of gedcom lines.
}
\description{
This function applies concatenation indicated by CONC/CONT lines.
}
