% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{create_conc_lines}
\alias{create_conc_lines}
\title{Create CONCatenation lines}
\usage{
create_conc_lines(lines, char_limit)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}

\item{char_limit}{Character limit of line values.}
}
\value{
A new character vector of gedcom lines, possibly expanded to include CONC lines.
}
\description{
Create CONCatenation lines
}
