% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{create_cont_lines}
\alias{create_cont_lines}
\title{Create CONTinuation lines}
\usage{
create_cont_lines(lines)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}
}
\value{
A new character vector of gedcom lines, possibly expanded to include CONT lines.
}
\description{
Create CONTinuation lines
}
