% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_change_date.R
\name{CreationDate}
\alias{CreationDate}
\title{Create a creation date object}
\usage{
CreationDate(date_exact = character(0), time = character(0))
}
\arguments{
\item{date_exact}{An exact date given either as a formatted GEDCOM string, or a
\code{DateExact()} object. If not given, it will default to today's date.}

\item{time}{The time given either as a formatted GEDCOM string, or a
\code{Time()} object.}
}
\value{
An S7 object representing a GEDCOM CREATION_DATE.
}
\description{
Create a creation date object
}
