% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{date_diff}
\alias{date_diff}
\title{Determine the number of years between two dates}
\usage{
date_diff(date1, date2 = date_exact_current()@c_as_val, minimise = TRUE)
}
\arguments{
\item{date1}{A GEDCOM date string.}

\item{date2}{A GEDCOM date string. If no date is given, today's date is used.}

\item{minimise}{If date ranges or periods are used in the dates, whether to choose the bounds which
assume the minimum date difference. If this is FALSE, the maximum date difference is assumed.}
}
\value{
A numeric value giving the number of years. A numeric value less than zero means no
determination could be made.
}
\description{
Determine the number of years between two dates
}
\keyword{internal}
