% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_date.R
\name{date_exact_current}
\alias{date_exact_current}
\title{Create a GEDCOM Exact Date object for today}
\usage{
date_exact_current()
}
\value{
An S7 object representing a GEDCOM Exact Date for today.
}
\description{
Create a GEDCOM Exact Date object for today
}
\keyword{internal}
