% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_date.R
\name{DateApprox}
\alias{DateApprox}
\title{Create a GEDCOM Approximate Date object}
\usage{
DateApprox(date_greg = character(0), about = TRUE, calc = FALSE, est = FALSE)
}
\arguments{
\item{date_greg}{A Gregorian date given either as a formatted GEDCOM string, or a
\code{DateGregorian} object.}

\item{about}{Whether the date is near to the date given.}

\item{calc}{Whether the date is calculated from other values.}

\item{est}{Whether the date is near to the date given, and is calculated from other values.}
}
\value{
An S7 object representing a GEDCOM Approximate Date.
}
\description{
Create a GEDCOM Approximate Date object
}
