% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_base_dates.R
\name{DateCalendar}
\alias{DateCalendar}
\title{Create a GEDCOM Calendar Date object}
\usage{
DateCalendar(
  year = integer(0),
  month = integer(0),
  day = integer(0),
  julian = FALSE
)
}
\arguments{
\item{year}{The year given as an integer (not 0). Negative years are interpreted
as Before the Common Era (BCE). If BCE, only year should be provided.}

\item{month}{The month of the year given as an integer between 1 and 12.}

\item{day}{The day of the month given as an integer between 1 and 31.}

\item{julian}{Whether the date is given in the Julian calendar. If not, it is
assumed to be in the Gregorian calendar.}
}
\value{
An S7 object representing a GEDCOM Calendar Date.
}
\description{
Create a GEDCOM Calendar Date object
}
