% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_date.R
\name{DateExact}
\alias{DateExact}
\title{Create a GEDCOM Exact Date object}
\usage{
DateExact(year = integer(0), month = integer(0), day = integer(0))
}
\arguments{
\item{year}{The year given as an integer (greater than 0).}

\item{month}{The month of the year given as an integer between 1 and 12.}

\item{day}{The day of the month given as an integer between 1 and 31.}
}
\value{
An S7 object representing a GEDCOM Exact Date.
}
\description{
Create a GEDCOM Exact Date object
}
