% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_base_dates.R
\name{DatePeriod}
\alias{DatePeriod}
\title{Create a GEDCOM Date Period object}
\usage{
DatePeriod(start_date = character(0), end_date = character(0))
}
\arguments{
\item{start_date}{The start of the period/range given either as a formatted GEDCOM string, or a
\code{DateGregorian} object.}

\item{end_date}{The end of the period/range given either as a formatted GEDCOM string, or a
\code{DateGregorian} object.}
}
\value{
An S7 object representing a GEDCOM Date Period.
}
\description{
Create a GEDCOM Date Period object
}
