% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_base_dates.R
\name{DateRange}
\alias{DateRange}
\title{Create a GEDCOM Date Range object}
\usage{
DateRange(start_date = character(0), end_date = character(0))
}
\arguments{
\item{start_date}{The start of the period/range given either as a formatted GEDCOM string, or a
\code{DateCalendar()} object.}

\item{end_date}{The end of the period/range given either as a formatted GEDCOM string, or a
\code{DateCalendar()} object.}
}
\value{
An S7 object representing a GEDCOM Date Range.
}
\description{
Create a GEDCOM Date Range object
}
