% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{delete_ged_section}
\alias{delete_ged_section}
\title{Delete a structure from GEDCOM lines}
\usage{
delete_ged_section(lines, line_no, containing_line = TRUE)
}
\arguments{
\item{lines}{A character vector of GEDCOM lines.}

\item{line_no}{A line number where the structure is located.}

\item{containing_line}{Whether the line number is the first line of the structure or
whether the line number references a line within the structure (but not more than one
level lower).}
}
\value{
The character vector of GEDCOM lines without the structure referenced by
the line_no. If the structure is an entire record, then any xref pointers to it
will also be replaced with a VOID pointer.
}
\description{
Delete a structure from GEDCOM lines
}
