% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{delete_ged_sections}
\alias{delete_ged_sections}
\title{Delete potentially multiple structures from GEDCOM lines}
\usage{
delete_ged_sections(lines, line_fn, containing_line = TRUE)
}
\arguments{
\item{lines}{A character vector of GEDCOM lines.}

\item{line_fn}{A callback function which takes a single input, lines, and
returns an integer vector identifying the row of each structure to be deleted.}

\item{containing_line}{Whether the line number(s) returned by the callback
are the first line of the structure(s) or whether the line number(s)
reference a line within the structure (but not more than one
level lower).}
}
\value{
The character vector of GEDCOM lines without the structures
identified by the callback function.
}
\description{
Delete potentially multiple structures from GEDCOM lines
}
\keyword{internal}
