% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_fam}
\alias{df_fam}
\title{Summarise Family records in a dataframe}
\usage{
df_fam(x, xrefs = NULL)
}
\arguments{
\item{x}{A gedcom object.}

\item{xrefs}{A vector of xrefs to summarise. If this is left NULL,
all relevant xrefs will be used.}
}
\value{
A dataframe summarising a Family record on each row.
}
\description{
Summarise Family records in a dataframe
}
