% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_fam_facts}
\alias{df_fam_facts}
\title{Summarise a family's attributes/events in a dataframe}
\usage{
df_fam_facts(x, xref)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The cross-reference identifier of a family record.}
}
\value{
A dataframe summarising an attribute/event on each row.
}
\description{
Summarise a family's attributes/events in a dataframe
}
