% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_indi}
\alias{df_indi}
\alias{df_fam}
\alias{df_sour}
\alias{df_repo}
\alias{df_media}
\alias{df_note}
\alias{df_subm}
\title{Summarise records of a particular type in a dataframe}
\usage{
df_indi(x)

df_fam(x)

df_sour(x)

df_repo(x)

df_media(x)

df_note(x)

df_subm(x)
}
\arguments{
\item{x}{A gedcom object.}
}
\value{
A dataframe summarising a record on each row.
}
\description{
Summarise records of a particular type in a dataframe
}
