% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_indi_facts}
\alias{df_indi_facts}
\title{Summarise an individual's attributes/events in a dataframe}
\usage{
df_indi_facts(x, xref)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The cross-reference identifier of an individual record.}
}
\value{
A dataframe summarising an attribute/event on each row.
}
\description{
Summarise an individual's attributes/events in a dataframe
}
