% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_media}
\alias{df_media}
\title{Summarise Multimedia records in a dataframe}
\usage{
df_media(x, xrefs = NULL)
}
\arguments{
\item{x}{A gedcom object.}

\item{xrefs}{A vector of xrefs to summarise. If this is left NULL,
all relevant xrefs will be used.}
}
\value{
A dataframe summarising a Multimedia record on each row.
}
\description{
Summarise Multimedia records in a dataframe
}
