% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_note}
\alias{df_note}
\title{Summarise Note records in a dataframe}
\usage{
df_note(x, xrefs = NULL)
}
\arguments{
\item{x}{A gedcom object.}

\item{xrefs}{A vector of xrefs to summarise. If this is left NULL,
all relevant xrefs will be used.}
}
\value{
A dataframe summarising a Note record on each row.
}
\description{
Summarise Note records in a dataframe
}
