% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_repo}
\alias{df_repo}
\title{Summarise Repository records in a dataframe}
\usage{
df_repo(x, xrefs = NULL)
}
\arguments{
\item{x}{A gedcom object.}

\item{xrefs}{A vector of xrefs to summarise. If this is left NULL,
all relevant xrefs will be used.}
}
\value{
A dataframe summarising a Repository record on each row.
}
\description{
Summarise Repository records in a dataframe
}
