% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_df.R
\name{df_sour}
\alias{df_sour}
\title{Summarise Source records in a dataframe}
\usage{
df_sour(x, xrefs = NULL)
}
\arguments{
\item{x}{A gedcom object.}

\item{xrefs}{A vector of xrefs to summarise. If this is left NULL,
all relevant xrefs will be used.}
}
\value{
A dataframe summarising a Source record on each row.
}
\description{
Summarise Source records in a dataframe
}
