% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_fact.R
\name{fact_rules_df}
\alias{fact_rules_df}
\title{Property requirements for fact objects.}
\usage{
fact_rules_df()
}
\value{
A dataframe detailing the property requirements for each type of fact.
}
\description{
Property requirements for fact objects.
}
