% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{FactsRecorded}
\alias{FactsRecorded}
\title{Create an object recording facts covered in a source record}
\usage{
FactsRecorded(
  fact_types = character(0),
  date_period = character(0),
  date_phrase = character(0),
  territory = NULL
)
}
\arguments{
\item{fact_types}{A character string indicating the types
of events that were recorded in a particular source. Each event type is separated by a
comma and space. For example, a parish register of births, deaths, and marriages
would be BIRT, DEAT, MARR.}

\item{date_period}{A date period given either as a formatted GEDCOM string, or a
\code{DatePeriod()} object.}

\item{date_phrase}{Textual information that cannot be expressed in the date.}

\item{territory}{The territory associated with the events covered. This can either be a
\code{Place()} object or a character string (a comma-separated string of region names,
ordered from smallest to largest).}
}
\value{
An S7 object representing a GEDCOM SOUR.EVEN structure.
}
\description{
Create an object recording facts covered in a source record
}
