% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_fact_fam.R
\name{FamilyEvent}
\alias{FamilyEvent}
\title{Create a family event object}
\usage{
FamilyEvent(
  fact_type = character(0),
  fact_val = character(0),
  fact_desc = character(0),
  date = character(0),
  place = NULL,
  address = NULL,
  phone_numbers = character(0),
  emails = character(0),
  faxes = character(0),
  web_pages = character(0),
  agency = character(0),
  relig_affil = character(0),
  cause = character(0),
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  date_sort = character(0),
  associations = list(),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  unique_ids = character(0),
  husb_age = character(0),
  husb_age_phrase = character(0),
  wife_age = character(0),
  wife_age_phrase = character(0)
)
}
\arguments{
\item{fact_type}{A code indicating the type of fact. This must be taken from one of
\code{val_individual_event_types()}, \code{val_individual_attribute_types()},
\code{val_family_event_types()}, or \code{val_family_attribute_types()}. A generic event ("EVEN")
or attribute ("FACT") can also be defined for more bespoke facts.}

\item{fact_val}{A value associated with the fact. For example for "NCHI" this would
be the number of children. This must be given as a character string.}

\item{fact_desc}{TODO}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{DateValue()} object.}

\item{place}{see territory}

\item{address}{The address given either as a \code{Address()} object or as a character string.
This would be as written on a mailing label with new lines separated by \\n.}

\item{phone_numbers}{A character vector of phone numbers.}

\item{emails}{A character vector of email addresses.}

\item{faxes}{A character vector of fax numbers.}

\item{web_pages}{A character vector of web page URLs.}

\item{agency}{The organization, institution, corporation, person, or other entity that
has responsibility for the associated fact. Examples are an employer of a person of an
associated occupation, or an educational establishment that has awarded a scholastic award.}

\item{relig_affil}{A religious denomination associated with the fact.}

\item{cause}{Used in special cases to record the reasons which precipitated the fact (e.g. cause of death).}

\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{date_sort}{The date given either as a formatted GEDCOM string, or a
\code{DateSorting()} object.}

\item{associations}{Associated individuals. This can either be a \code{Association()} object or a list of them.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Associated multimedia. This can either be a \code{MediaLink()} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{husb_age}{A character string that indicates the age in years, months, weeks and/or days
that the husband was at the time of the fact. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 1w 3d".
Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age
in the corresponding phrase parameter.}

\item{husb_age_phrase}{Free text information that cannot be expressed in the husband's age.}

\item{wife_age}{A character string that indicates the age in years, months, weeks and/or days
that the wife was at the time of the fact. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 1w 3d".
Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age
in the corresponding phrase parameter.}

\item{wife_age_phrase}{Free text information that cannot be expressed in the wife's age.}
}
\value{
An S7 object representing a GEDCOM FAMILY_EVENT_STRUCTURE.
}
\description{
Create a family event object
}
