% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_family_link.R
\name{FamilyLinkChild}
\alias{FamilyLinkChild}
\title{Create a family link (as child) object}
\usage{
FamilyLinkChild(
  fam_xref = character(0),
  note_xrefs = character(0),
  notes = list(),
  pedigree = character(0),
  pedigree_phrase = character(0),
  confidence = character(0),
  confidence_phrase = character(0)
)
}
\arguments{
\item{fam_xref}{The cross-reference identifier of a family record.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes.}

\item{pedigree}{An optional value from \code{val_pedigree_types()} indicating the nature
of the link.}

\item{pedigree_phrase}{An optional free-text phrase describing the nature of the
link. This is required if \verb{@pedigree} is "OTHER".}

\item{confidence}{An optional value from \code{val_confidence_types()} indicating the
confidence of the link.}

\item{confidence_phrase}{An optional free-text phrase expanding on the confidence of the
link.}
}
\value{
An S7 object representing a GEDCOM family link as a child.
}
\description{
Create a family link (as child) object
}
