% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_fam.R
\name{FamilyRecord}
\alias{FamilyRecord}
\title{Create a family record object}
\usage{
FamilyRecord(
  xref = "@GEDCOMS7_ORPHAN@",
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  user_ids = character(0),
  unique_ids = character(0),
  ext_ids = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  created = NULL,
  updated = NULL,
  facts = list(),
  non_events = list(),
  husb_xref = character(0),
  wife_xref = character(0),
  chil_xrefs = character(0),
  associations = list(),
  subm_xrefs = character(0),
  spouse_sealings = list()
)
}
\arguments{
\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{user_ids}{A character vector of user-generated identifiers. The type of the identifiers can
be given in the vector names, e.g. c("Driving license number" = "ABC123")}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{A named character vector of identifiers maintained by an external authority.
The names must be given as a URI. See the GEDCOM specification for more information.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Associated multimedia. This can either be a \code{MediaLink()} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{created}{A \code{CreationDate()} object containing the date the record was created.
Creating an object with no parameters sets the date to today.}

\item{updated}{A \code{ChangeDate()} object containing the date the record was updated.
Creating an object with no parameters sets the date to today.}

\item{facts}{TODO}

\item{non_events}{TODO}

\item{husb_xref, wife_xref, chil_xrefs}{The cross-reference identifier(s) of member's individual records.
If the individual does not have a record, then the value "@VOID@" can be used.
However, you will need to describe the individual by using a named vector (a description can be used
in either case), e.g. c("Joe Bloggs" = "@VOID@") or c("Joe Bloggs" = "@I1@").}

\item{associations}{Associated individuals. This can either be a \code{Association()} object or a list of them.}

\item{subm_xrefs}{A character vector of relevant submitter record cross-reference identifiers.}

\item{spouse_sealings}{TODO}
}
\value{
An S7 object representing a GEDCOM FAMILY_RECORD.
}
\description{
Create a family record object
}
