% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{GedcomHeader}
\alias{GedcomHeader}
\title{Create a GEDCOM header object}
\usage{
GedcomHeader(
  gedcom_version = character(0),
  ext_tags = character(0),
  source = NULL,
  destination = character(0),
  creation_date = <object>,
  creation_time = character(0),
  subm_xref = character(0),
  gedcom_copyright = character(0),
  default_language = character(0),
  default_place_form = character(0),
  notes = list(),
  note_xrefs = character(0)
)
}
\arguments{
\item{gedcom_version}{TODO}

\item{ext_tags}{TODO}

\item{source}{TODO}

\item{destination}{TODO}

\item{creation_date}{The creation date of the file given either as a formatted GEDCOM string, or a
\code{DateExact} object.}

\item{creation_time}{The creation time of the file given either as a formatted GEDCOM string, or a
\code{Time} object.}

\item{subm_xref}{The cross-reference identifier of a submitter record.}

\item{gedcom_copyright}{TODO}

\item{default_language}{TODO}

\item{default_place_form}{TODO}

\item{notes}{Associated notes. This can either be a \code{Note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}
}
\value{
An S7 object representing a GEDCOM HEADER.
}
\description{
Create a GEDCOM header object
}
