% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{GedcomS7}
\alias{GedcomS7}
\title{Create a GEDCOM object}
\usage{
GedcomS7(
  header = GedcomHeader(),
  records = GedcomRecords(),
  update_change_dates = FALSE,
  add_creation_dates = FALSE
)
}
\arguments{
\item{records}{An S7 object whose properties contain information about all records. Do not edit
properties in capitals directly. See Details for more information.}

\item{update_change_dates}{Whether to automatically update change dates when updating records.
This happens when the record is pushed to the gedcom object.}

\item{add_creation_dates}{Whether to automatically add creation dates when creating records.
This happens when the record is pushed to the gedcom object.}

\item{gedcom_version}{The version number of the official specification that this
GEDCOM conforms to. This must include the major and minor version (for example, “7.0”);
it may include the patch as well (for example, “7.0.1”), but doing so is not required.}

\item{ext_tags}{Not supported.}

\item{source}{A \code{GedcomSource()} object describing the software that has generated the GEDCOM.`}

\item{destination}{An identifier for the system expected to receive this GEDCOM.}

\item{creation_date}{The creation date of the file given either as a formatted GEDCOM string, or a
\code{DateExact()} object.}

\item{creation_time}{The creation time of the file given either as a formatted GEDCOM string, or a
\code{Time()} object.}

\item{subm_xref}{The cross-reference identifier of the primary submitter.}

\item{gedcom_copyright}{A copyright statement, as appropriate for the copyright laws applicable to
this data.}

\item{default_language}{The default language for the entire GEDCOM object.}

\item{default_place_form}{The default form for place names in the GEDCOM object.
A comma-separated string of jurisdictional titles.
For example "City, County, State, Country".}
}
\value{
An S7 object representing a GEDCOM file.
}
\description{
You shouldn't need to use this directly to create new GEDCOM objects. Instead, use
\code{new_gedcom()} which populates relevant defaults.
}
\details{
All information about records is contained in the \verb{@records} property.

The \verb{@prefixes} property is a named vector containing any alphanumeric string (up to 6 characters long)
which will precede the number given to identify new records, of which there are currently 7 types:

Individual (INDI)
Family (FAM)
Source (SOUR)
Repository (REPO)
Multimedia (OBJE)
Note (SNOTE)
Submitter (SUBM)

This vector must be of a particular length with specific names. For example:
c(SUBM = "U", INDI = "I", FAM = "F", SOUR = "S", REPO = "R", OBJE = "M", SNOTE = "N").

The order that these records appear in the vector will also dictate the order in which records
will appear in the exported file.
}
