% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{GedcomSource}
\alias{GedcomSource}
\title{Create a GEDCOM source object}
\usage{
GedcomSource(
  product_id = character(0),
  product_name = character(0),
  product_version = character(0),
  business_name = character(0),
  business_address = NULL,
  phone_numbers = character(0),
  emails = character(0),
  faxes = character(0),
  web_pages = character(0),
  data_name = character(0),
  data_pubdate = character(0),
  data_pubtime = character(0),
  data_copyright = character(0)
)
}
\arguments{
\item{product_id}{An identifier for the product producing this dataset.}

\item{product_name}{The name of the product producing this dataset.}

\item{product_version}{The version of the product producing this dataset.}

\item{business_name}{The name of the business, corporation, or person that produced
or commissioned the product.}

\item{business_address}{The address of the business, corporation, or person that produced
or commissioned the product. The address is given either as a \code{Address()} object or as a
character string. This would be as written on a mailing label with new lines separated by \\n.}

\item{phone_numbers}{A character vector of phone numbers.}

\item{emails}{A character vector of email addresses.}

\item{faxes}{A character vector of fax numbers.}

\item{web_pages}{A character vector of web page URLs.}

\item{data_name}{Deprecated.}

\item{data_pubdate}{Deprecated.}

\item{data_pubtime}{Deprecated.}

\item{data_copyright}{Deprecated.}
}
\value{
An S7 object representing a GEDCOM HEAD.SOUR.
}
\description{
Create a GEDCOM source object
}
