% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xrefs.R
\name{get_descendants}
\alias{get_descendants}
\title{Identify all descendants for an individual}
\usage{
get_descendants(
  x,
  xref,
  inc_indi = FALSE,
  inc_part = FALSE,
  inc_famg = FALSE,
  inc_supp = FALSE,
  birth_only = TRUE
)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{inc_indi}{Whether to also include the individual themselves.}

\item{inc_part}{Whether to also include all partners of this individual (and their descendants and
descendants' partners).}

\item{inc_famg}{Whether to also include all Family Group records where this individual is a partner
(and all descendants' Family Group records).}

\item{inc_supp}{Whether to also include all supporting records (Note, Source, Repository, Multimedia).}

\item{birth_only}{Whether to only include biological descendants.}
}
\value{
A vector of xrefs of descendants.
}
\description{
This function identifies records in an entire branch of the family tree below a certain individual.
}
