% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fam_as_partner}
\alias{get_fam_as_partner}
\title{Identify all families for an individual where they are a partner}
\usage{
get_fam_as_partner(x, xref)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}
}
\value{
A character vector of xrefs.
}
\description{
Identify all families for an individual where they are a partner
}
