% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_famg_as_child}
\alias{get_famg_as_child}
\title{Identify all families for an individual where they are a child}
\usage{
get_famg_as_child(x, xref, birth_only = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{birth_only}{Whether to only return the family containing the biological parents.}
}
\value{
A character vector of family xrefs.
}
\description{
Identify all families for an individual where they are a child
}
