% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xrefs.R
\name{get_famg_children}
\alias{get_famg_children}
\title{Identify all children in a Family Group}
\usage{
get_famg_children(x, xref, birth_only = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of a Family Group record.}

\item{birth_only}{Whether to only return biological children.}
}
\value{
A character vector of children xrefs.
}
\description{
Identify all children in a Family Group
}
