% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_famg_partners}
\alias{get_famg_partners}
\title{Identify all partners in a Family Group}
\usage{
get_famg_partners(x, xref)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of a Family Group record.}
}
\value{
A character vector of partner xrefs.
}
\description{
Identify all partners in a Family Group
}
