% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_children}
\alias{get_indi_children}
\title{Identify all children for an individual}
\usage{
get_indi_children(x, xref, birth_only = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{birth_only}{Whether to only return biological children.}
}
\value{
A character vector of children xrefs.
}
\description{
Identify all children for an individual
}
