% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_cousins}
\alias{get_indi_cousins}
\title{Identify all cousins for an individual}
\usage{
get_indi_cousins(x, xref, degree = 1, inc_half = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{degree}{Whether to return first cousins (degree = 1), second cousins (degree = 2), etc.}

\item{inc_half}{Whether to include half cousins.}
}
\value{
A character vector of cousin xrefs.
}
\description{
Identify all cousins for an individual
}
