% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_fathers}
\alias{get_indi_fathers}
\title{Identify all fathers for an individual}
\usage{
get_indi_fathers(x, xref, birth_only = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{birth_only}{Whether to only return the biological father.}
}
\value{
A character vector of father xrefs.
}
\description{
Identify all fathers for an individual
}
