% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_mothers}
\alias{get_indi_mothers}
\title{Identify all mothers for an individual}
\usage{
get_indi_mothers(x, xref, birth_only = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{birth_only}{Whether to only return the biological mother.}
}
\value{
A character vector of mother xrefs.
}
\description{
Identify all mothers for an individual
}
