% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_siblings}
\alias{get_indi_siblings}
\title{Identify all siblings for an individual}
\usage{
get_indi_siblings(x, xref, pedigrees = NULL, inc_half = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{pedigrees}{A character vector of allowed family-child linkages.
By default, NULL means all pedigrees (e.g. inc. ADOPTED).
If it includes "BIRTH" then this will also pick up non-existent values
(since BIRTH is assumed to be the default).}

\item{inc_half}{Whether to include siblings that only share one parent.}
}
\value{
A character vector of xrefs.
}
\description{
Identify all siblings for an individual
}
