% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_records_by_pattern}
\alias{get_records_by_pattern}
\title{Identify all records that contain a pattern}
\usage{
get_records_by_pattern(x, pattern, return_context = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{pattern}{A regular expression. Case is ignored.}

\item{return_context}{Whether to return a named list of matching lines in
the GEDCOM (TRUE) or just the xrefs (FALSE, default).}
}
\value{
Either a vector of matching xrefs, or a named list of matching GEDCOM lines.
}
\description{
Identify all records that contain a pattern
}
