% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_supporting_recs}
\alias{get_supporting_recs}
\title{Identify all supporting records for a set of records}
\usage{
get_supporting_recs(
  x,
  xrefs,
  inc_note = TRUE,
  inc_media = TRUE,
  inc_sour = TRUE,
  inc_repo = TRUE
)
}
\arguments{
\item{x}{A gedcom object.}

\item{xrefs}{The xrefs of records to get supporting records for.}

\item{inc_note}{Whether to include Note records.}

\item{inc_media}{Whether to include Multimedia records.}

\item{inc_sour}{Whether to include Source records.}

\item{inc_repo}{Whether to include Repository records.}
}
\value{
A character vector of xrefs.
}
\description{
This function gets all supporting records (and onwards dependencies) for a set of records. Supporting records
are note, multimedia, source, and repository records, i.e. those providing supporting evidence and comments.
}
