% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_unused_recs}
\alias{get_unused_recs}
\title{Identify unreferenced records}
\usage{
get_unused_recs(x)
}
\arguments{
\item{x}{A gedcom object.}
}
\value{
A character vector of xrefs.
}
\description{
This function identifies records that are not referenced in any other records.
}
\details{
You would expect every record to be referenced by another in some way. For example, Individual
records should reference Family Group records (and vice-versa), Repository records should be referenced
by Source records, and Source records should be cited by other records.
}
