% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{guess_age_from_famg_events}
\alias{guess_age_from_famg_events}
\title{Guess an individual's age from their family group events}
\usage{
guess_age_from_famg_events(tg, xref, agg_fn = mean)
}
\arguments{
\item{tg}{A tidyged object.}

\item{xref}{The xref of an individual.}

\item{agg_fn}{If multiple ages are calculated, the function to use to aggregate them.}
}
\value{
A numeric value giving the estimated age. A numeric value less than zero means no
determination could be made.
}
\description{
This function takes an individual's family group events and calculates an estimated age based on the date of the event and their age when the event occurred.
}
