% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{increase_level}
\alias{increase_level}
\title{Increase the level of a vector of GEDCOM lines}
\usage{
increase_level(ged, by = 1)
}
\arguments{
\item{ged}{A character vector of GEDCOM lines.}

\item{by}{The number of levels to increment.}
}
\value{
The vector of GEDCOM lines with incremented levels.
}
\description{
Increase the level of a vector of GEDCOM lines
}
