% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_indi.R
\name{IndividualRecord}
\alias{IndividualRecord}
\title{Create an individual record object}
\usage{
IndividualRecord(
  xref = "@GEDCOMS7_ORPHAN@",
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  user_ids = character(0),
  unique_ids = character(0),
  ext_ids = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  created = NULL,
  updated = NULL,
  pers_names = list(),
  sex = "U",
  facts = list(),
  non_events = list(),
  ordinances = list(),
  fam_links_chil = list(),
  fam_links_spou = list(),
  subm_xrefs = character(0),
  associations = list(),
  alia_xrefs = character(0),
  anci_xrefs = character(0),
  desi_xrefs = character(0)
)
}
\arguments{
\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{user_ids}{A character vector of user-generated identifiers. The type of the identifiers can
be given in the vector names, e.g. c("Driving license number" = "ABC123")}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{A named character vector of identifiers maintained by an external authority.
The names must be given as a URI. See the GEDCOM specification for more information.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Associated multimedia. This can either be a \code{MediaLink()} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{created}{A \code{CreationDate()} object containing the date the record was created.
Creating an object with no parameters sets the date to today.}

\item{updated}{A \code{ChangeDate()} object containing the date the record was updated.
Creating an object with no parameters sets the date to today.}

\item{pers_names}{A \code{PersonalName()} object or a list of them, giving the names
of this individual. A simple character vector of names can be provided instead,
but this is not recommended.}

\item{sex}{The sex of the individual. Either "M" (male), "F" (female), "X" (other), or
"U" (undetermined, the default).}

\item{facts}{Events and/or attributes for this individual.
An \code{IndividualEvent()}/\code{IndividualAttribute()} object, or a list of them.}

\item{non_events}{Events that this individual did not experience.
A \code{NonEvent()} object, or a list of them.}

\item{ordinances}{An \code{Ordinance()} object, or a list of them.}

\item{fam_links_chil}{A \code{FamilyLinkChild()} object or a list of them, giving the
Family records that this individual is a member of as a child. A character vector
of Family record xrefs can also be provided. This will be automatically updated if
the individual's membership in a Family record changes.}

\item{fam_links_spou}{A \code{FamilyLinkSpouse()} object or a list of them, giving the
Family records that this individual is a member of as a spouse. A character vector
of Family record xrefs can also be provided. This will be automatically updated if
the individual's membership in a Family record changes.}

\item{subm_xrefs}{A character vector of relevant submitter record cross-reference identifiers.}

\item{associations}{Associated individuals. This can either be a \code{Association()} object or a list of them.}

\item{alia_xrefs}{A named character vector of relevant individual record cross-reference identifiers
whose records also represent this individual. The vector names may provide a description of these records.}

\item{anci_xrefs}{A character vector of relevant submitter record cross-reference identifiers
who are interested in the ancestors of this individual.}

\item{desi_xrefs}{A character vector of relevant submitter record cross-reference identifiers
who are interested in the descendants of this individual.}
}
\value{
An S7 object representing a GEDCOM INDIVIDUAL_RECORD.
}
\description{
Create an individual record object
}
