% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_media.R
\name{MediaFile}
\alias{MediaFile}
\title{Create a media file object}
\usage{
MediaFile(
  location = character(0),
  title = character(0),
  media_type = character(0),
  medium = character(0),
  medium_phrase = character(0),
  media_alt = character(0)
)
}
\arguments{
\item{location}{An absolute or relative URL to the file.}

\item{title}{The title of the multimedia record.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{medium}{A value from \code{val_medium_types()}. If "OTHER" is selected then a \verb{@medium_phrase}
must be given. This should describe the original medium from which it was derived. So
if it is a digital image scanned from a physical photograph, it should be "PHOTO"
instead of "ELECTRONIC".}

\item{medium_phrase}{A free text description of the medium. This is mandatory if \verb{@medium} is
"OTHER".}

\item{media_alt}{A named vector of the media in alternative media forms, c(form = location)}
}
\value{
An S7 object representing a GEDCOM multimedia file substructure.
}
\description{
Create a media file object
}
