% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_media.R
\name{MediaRecord}
\alias{MediaRecord}
\title{Create a multimedia record object}
\usage{
MediaRecord(
  xref = "@GEDCOMS7_ORPHAN@",
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  user_ids = character(0),
  unique_ids = character(0),
  ext_ids = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  created = NULL,
  updated = NULL,
  files = list()
)
}
\arguments{
\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{user_ids}{A character vector of user-generated identifiers. The type of the identifiers can
be given in the vector names, e.g. c("Driving license number" = "ABC123")}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{A named character vector of identifiers maintained by an external authority.
The names must be given as a URI. See the GEDCOM specification for more information.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records.}

\item{media_links}{Not used.}

\item{created}{A \code{CreationDate()} object containing the date the record was created.
Creating an object with no parameters sets the date to today.}

\item{updated}{A \code{ChangeDate()} object containing the date the record was updated.
Creating an object with no parameters sets the date to today.}

\item{files}{A \code{MediaFile} object or a list of them. This refers to 1 or more external
digital files. Grouped files should each pertain to the same context.}
}
\value{
An S7 object representing a GEDCOM MULTIMEDIA_RECORD.
}
\description{
Create a multimedia record object
}
