% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{new_gedcom}
\alias{new_gedcom}
\title{Create a new GEDCOM object}
\usage{
new_gedcom(my_language = "en")
}
\arguments{
\item{my_language}{The primary language in which data will be stored. The language code should
adhere to BCP 47.}
}
\value{
A minimal gedcom S7 object.
}
\description{
Create a new GEDCOM object
}
\details{
See documentation for \code{GedcomS7()} and \code{GedcomSource()} for object properties.
}
