% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{new_gedcomR7}
\alias{new_gedcomR7}
\title{Create a new gedcom object}
\usage{
new_gedcomR7(my_name = unname(Sys.info()["user"]), my_language = "English")
}
\arguments{
\item{my_name}{Your name, as creator/submitter.}

\item{my_language}{The primary language in which data will be stored.}
}
\value{
A minimal gedcom R7 object.
}
\description{
Create a new gedcom object
}
