% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_non_event.R
\name{NonEvent}
\alias{NonEvent}
\title{Create a non-event object}
\usage{
NonEvent(
  event_type = character(0),
  date_period = character(0),
  date_phrase = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list()
)
}
\arguments{
\item{event_type}{TODO}

\item{date_period}{A date period given either as a formatted GEDCOM string, or a
\code{DatePeriod()} object.}

\item{date_phrase}{Textual information that cannot be expressed in the date.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records.}
}
\value{
An S7 object representing a GEDCOM NON_EVENT_STRUCTURE.
}
\description{
Create a non-event object
}
