% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_note.R
\name{NoteRecord}
\alias{NoteRecord}
\title{Create a shared note record object}
\usage{
NoteRecord(
  XREF = "@GEDCOMS7_ORPHAN@",
  confidential = FALSE,
  locked = FALSE,
  private = FALSE,
  user_ids = character(0),
  unique_ids = character(0),
  ext_ids = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  media_links = list(),
  created = NULL,
  updated = NULL,
  text = character(0),
  media_type = character(0),
  language = character(0),
  translations = list()
)
}
\arguments{
\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{user_ids}{A character vector of user-generated identifiers. The type of the identifiers can
be given in the vector names, e.g. c("Driving license number" = "ABC123")}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{A named character vector of identifiers maintained by an external authority.
The names must be given as a URI. See the GEDCOM specification for more information.}

\item{note_xrefs}{Not used.}

\item{notes}{Not used.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records.}

\item{media_links}{Not used.}

\item{created}{A \code{CreationDate()} object containing the date the record was created.
Creating an object with no parameters sets the date to today.}

\item{updated}{A \code{ChangeDate()} object containing the date the record was updated.
Creating an object with no parameters sets the date to today.}

\item{text}{A character string. New lines are created with \\n.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{language}{A character string of language tags as defined in BCP 47.}

\item{translations}{A \code{TranslationText()} object or a list of them.
One for each alternate translation of the text.}
}
\value{
An S7 object representing a GEDCOM SHARED_NOTE_RECORD.
}
\description{
Create a shared note record object
}
