% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{obj_to_ged}
\alias{obj_to_ged}
\title{Convert an input into a vector of GEDCOM lines}
\usage{
obj_to_ged(obj, tag = NULL)
}
\arguments{
\item{obj}{Either an atomic vector, S7 class object, or list.
Any S7 class objects must have an \code{as_ged()} method.}

\item{tag}{If the obj contains any atomic elements, then this
will specify what tag they are recorded against.}
}
\description{
Convert an input into a vector of GEDCOM lines
}
