% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ordinance.R
\name{Ordinance}
\alias{Ordinance}
\title{Create an individual ordinance object}
\usage{
Ordinance(
  ord_type = character(0),
  date = character(0),
  temple_name = character(0),
  place = character(0),
  ord_state = character(0),
  state_date = character(0),
  state_time = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list(),
  fam_xref = character(0)
)
}
\arguments{
\item{ord_type}{A value from \code{val_individual_ordinance_types()}.}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{DateValue()} object.}

\item{temple_name}{The name of a temple of The Church of Jesus Christ of Latter-day Saints.}

\item{place}{see territory}

\item{ord_state}{An optional value from \verb{val_ordinance_states(@ord_type|"SLGS")}.}

\item{state_date}{The ordinance date given either as a formatted GEDCOM string, or a
\code{DateExact} object.}

\item{state_time}{The ordinance time given either as a formatted GEDCOM string, or a
\code{Time} object.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{fam_xref}{The cross-reference identifier of a family record.}
}
\value{
An S7 object representing a GEDCOM LDS_INDIVIDUAL_ORDINANCE.
}
\description{
Create an individual ordinance object
}
