% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record.R
\name{parse_common_record_elements}
\alias{parse_common_record_elements}
\title{Parse common elements into a record object}
\usage{
parse_common_record_elements(rec, rec_lines)
}
\arguments{
\item{rec}{An S7 record object.}

\item{rec_lines}{A character vector of lines of a GEDCOM record.}
}
\value{
The S7 record object with common elements added as properties.
}
\description{
Parse common elements into a record object
}
\keyword{internal}
