% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{parse_gedcom_date}
\alias{parse_gedcom_date}
\title{Convert a GEDCOM date into a date object}
\usage{
parse_gedcom_date(date_string, minimise = TRUE)
}
\arguments{
\item{date_string}{A Gregorian date string.}

\item{minimise}{Whether to fill in missing date pieces so that the date is minimised.
For example, if no month is given, January is used. If minimise = FALSE, December will be used.}
}
\value{
A date.
}
\description{
Convert a GEDCOM date into a date object
}
