% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_personal_name.R
\name{PersonalNameTran}
\alias{PersonalNameTran}
\title{Create a name translation object}
\usage{
PersonalNameTran(
  pers_name = character(0),
  language = character(0),
  name_pieces = NULL
)
}
\arguments{
\item{pers_name}{TODO}

\item{language}{A character string of language tags as defined in BCP 47.}

\item{name_pieces}{A \code{PersonalNamePieces()} object defining the pieces of the full name.}
}
\value{
An S7 object representing a GEDCOM personal name translation substructure.
}
\description{
Create a name translation object
}
