% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_properties.R
\name{prop_char}
\alias{prop_char}
\title{Create a character-based property}
\usage{
prop_char(
  min_size = NULL,
  max_size = NULL,
  min_char = NULL,
  max_char = NULL,
  choices = NULL,
  pattern = NULL,
  names_required = FALSE,
  default = NULL,
  casting_name = NULL,
  S7class_names = NULL
)
}
\arguments{
\item{min_size}{Minimum number of elements allowed.}

\item{max_size}{Maximum number of elements allowed.}

\item{min_char}{Minimum number of characters per element.}

\item{max_char}{Maximum number of characters per element.}

\item{choices}{Character vector of allowed values.}

\item{pattern}{Regular expression pattern that values must match.}

\item{names_required}{Logical indicating whether property values should have names.}

\item{default}{Default value for the property.}

\item{casting_name}{The name of the property if you want to explicitly cast it
to character type when being set (this allows users to provide other atomic types).}

\item{S7class_names}{Character vector of gedcomS7 class names that this property
could also take.}
}
\value{
An S7 property object.
}
\description{
Constructs a character property with optional constraints on size,
character length, pattern matching, and allowed choices.
}
\keyword{internal}
