% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_property_defns.R
\name{prop_definitions}
\alias{prop_definitions}
\title{Define properties in GEDCOM 7.0 specification}
\arguments{
\item{full}{A full address as it would appear on a mailing label, with lines separated
by \n. For example:
"The White House\n1600 Pennsylvania Avenue, N.W.\nWashington, DC 20500\nUnited States of America"}

\item{city}{Optional. The city component of the address.}

\item{state}{Optional. The state component of the address.}

\item{postal_code}{Optional. The postal code component of the address.}

\item{country}{Optional. The country component of the address.}

\item{text}{A character string. New lines are created with \n.}

\item{language}{Optional. Language tags as defined in BCP 47.}

\item{media_type}{Optional. The media type as defined in RFC 2045.}

\item{alt_text}{Optional. A \code{class_translation_txt} object or a list of them.
One for each alternate translation of the text.}

\item{media_uid}{The UID of a multimedia record.}

\item{title}{The title of the multimedia record. This will supersede any title given
in the record.}

\item{crop}{Whether to crop the multimedia to a specific area.}

\item{top}{The number of pixels to omit from the top side of the image.}

\item{left}{The number of pixels to omit from the left side of the image.}

\item{height}{The height in pixels of the cropped region.}

\item{width}{The width in pixels of the cropped region.}

\item{hour}{The hour of the day given as an integer between 0 and 23.}

\item{minute}{The minute of the hour given as an integer between 0 and 59.}

\item{second}{Optional. The second of the minute given as an integer between 0 and 59.}

\item{fraction}{Optional (if @second is provided). The fraction of the second given as
an integer.}

\item{utc}{Whether the time is in Coordinated Universal Time (UTC) (TRUE, the default) or
is in local time (FALSE).}

\item{year}{The year given as an integer (greater than 0).}

\item{month}{The month of the year given as an integer between 1 and 12.}

\item{day}{The day of the month given as an integer between 1 and 31.}

\item{bce}{Whether the date is Before the Common Era. This is FALSE by default,
but if TRUE, only the year should be given.}

\item{date_greg}{A Gregorian date given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{about}{Whether the date is near to the date given.}

\item{calc}{Whether the date is calculated from other values.}

\item{est}{Whether the date is near to the date given, and is calculated from other values.}

\item{start_date}{The start of the period/range given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{end_date}{The end of the period/range given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_greg} / \code{class_date_period} / \code{class_date_range} / \code{class_date_approx} object.}

\item{date_phrase}{Optional. Textual information that cannot be expressed in the date.}

\item{time}{Optional. The time given either as a formatted GEDCOM string, or a
\code{class_time} object.}
}
\description{
Define all properties used in S7 classes.
}
\details{
This empty function serves as a single location where all properties are defined,
mainly for efficiency and maintainability.
}
