% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_properties.R
\name{prop_S7list}
\alias{prop_S7list}
\title{Create a property for a list of gedcomS7 objects}
\usage{
prop_S7list(prop_name, S7_class)
}
\arguments{
\item{prop_name}{A character string specifying the name of the property.}

\item{S7_class}{The gedcomS7 class that each element of the list should conform to.}
}
\value{
An S7 property object.
}
\description{
Constructs a property that stores a list of objects of a specified gedcomS7 class.
}
\details{
If any elements of the list are of type character, then they will automatically
be converted into objects of the appropriate class. Any elements that cannot
be converted will result in the validator throwing an error.
}
