% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_properties.R
\name{prop_S7obj}
\alias{prop_S7obj}
\title{Create a property for a single gedcomS7 object}
\usage{
prop_S7obj(prop_name, S7_class)
}
\arguments{
\item{prop_name}{A character string specifying the name of the property.}

\item{S7_class}{The gedcomS7 class that the property value should conform to.}
}
\value{
An S7 property object.
}
\description{
Constructs a property that can store a single object of a specified gedcomS7 class.
}
\details{
The property will default to a value of NULL if no object is provided.
}
\keyword{internal}
