% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record.R
\name{pull_record}
\alias{pull_record}
\title{Pull a record from a GEDCOM object for editing}
\usage{
pull_record(x, xref)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of the record to pull.}
}
\value{
An S7 object representing the record.
}
\description{
Pull a record from a GEDCOM object for editing
}
\details{
The record is not removed from the gedcom object, rather a copy is taken.
}
