% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_import.R
\name{read_gedcom}
\alias{read_gedcom}
\title{Import a GEDCOM file}
\usage{
read_gedcom(filepath = NULL, lines = NULL)
}
\arguments{
\item{filepath}{The full filepath of the GEDCOM file. If this and the \code{lines}
parameter are both NULL then you will be prompted to select a file.}

\item{lines}{If the filepath is not provided then a character vector of
GEDCOM lines can be provided.}
}
\value{
A gedcom S7 object.
}
\description{
Imports a *.ged file and creates a gedcom object.
}
