% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{remove_living}
\alias{remove_living}
\title{Remove living individuals in a gedcom object}
\usage{
remove_living(x, max_age = 100)
}
\arguments{
\item{x}{A gedcom object.}

\item{max_age}{The maximum age to assume for a living person (if a date of birth is given).}
}
\value{
A gedcom object cleansed of information on living individuals.
}
\description{
Remove living individuals in a gedcom object
}
