% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{remove_sensitive_sections}
\alias{remove_sensitive_sections}
\title{Remove gedcom structures marked as sensitive}
\usage{
remove_sensitive_sections(lines, restriction)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}

\item{restriction}{Whether to remove structures marked as "CONFIDENTIAL" or "PRIVACY".}
}
\value{
A vector of sanitised GEDCOM lines.
}
\description{
Remove gedcom structures marked as sensitive
}
