% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{SourceCallNumber}
\alias{SourceCallNumber}
\title{Create a source call number object}
\usage{
SourceCallNumber(
  call_number = character(0),
  medium = character(0),
  medium_phrase = character(0)
)
}
\arguments{
\item{call_number}{The call number.}

\item{medium}{A value from \code{val_medium_types()}. If "OTHER" is selected then a \verb{@medium_phrase}
must be given.}

\item{medium_phrase}{A free text description of the medium. This is mandatory if \verb{@medium} is
"OTHER".}
}
\value{
An S7 object representing the CALN substructure of a GEDCOM
SOURCE_REPOSITORY_CITATION.
}
\description{
Create a source call number object
}
