% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_citation.R
\name{SourceCitation}
\alias{SourceCitation}
\title{Create a source citation object}
\usage{
SourceCitation(
  sour_xref = "@VOID@",
  where = character(0),
  date = character(0),
  source_text = list(),
  fact_type = character(0),
  fact_phrase = character(0),
  role = character(0),
  role_phrase = character(0),
  certainty = character(0),
  media_links = list(),
  note_xrefs = character(0),
  notes = list()
)
}
\arguments{
\item{sour_xref}{The cross-reference identifier of a source record. If the source
does not have a record, then this can be left blank and the value "@VOID@" will be used. However,
you should describe the source in @where.}

\item{where}{A specific location within the information referenced. For a published work, this could
include the volume of a multi-volume work and the page number or numbers. For a
periodical, it could include volume, issue, and page numbers. For a newspaper, it could
include a date, page number, and column number. For an unpublished source or micro‐
filmed works, this could be a film or sheet number, page number, or frame number. A
census record might have an enumerating district, page number, line number, dwelling
number, and family number.
It is recommended that the data in this field be formatted comma-separated with label:
value pairs}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{DateValue()} object.}

\item{source_text}{A verbatim copy of any description contained within the source. This can
either be a \code{TranslationText()} object, a list of them, or a character vector of text.
If any \code{TranslationText()} objects are provided, then you will be forced to define an
associated language or media type.}

\item{fact_type}{A code indicating the type of fact. This must be taken from one of
\code{val_individual_event_types()}, \code{val_individual_attribute_types()},
\code{val_family_event_types()}, or \code{val_family_attribute_types()}. A generic event ("EVEN")
or attribute ("FACT") can also be defined for more bespoke facts. See \code{val_facts_rules()}
for the set of rules surrounding the codes allowed.}

\item{fact_phrase}{Textual information that cannot be expressed in the fact type.}

\item{role}{What role this person played in this fact.}

\item{role_phrase}{Textual information that cannot be expressed in the role.}

\item{certainty}{An enumerated value indicating the credibility of a
piece of information, based on its supporting evidence. Some systems use this feature to
rank multiple conflicting opinions for display of most likely information first. It is not
intended to eliminate the receivers’ need to evaluate the evidence for themselves.
"0" = unreliable/estimated data
"1" = Questionable reliability of evidence
"2" = Secondary evidence, data officially recorded sometime after event
"3" = Direct and primary evidence used, or by dominance of the evidence}

\item{media_links}{Associated multimedia. This can either be a \code{MediaLink()} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}
}
\value{
An S7 object representing a GEDCOM SOURCE_CITATION.
}
\description{
Create a source citation object
}
