% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{split_gedcom_values}
\alias{split_gedcom_values}
\title{Convert the GEDCOM form to GEDCOM grammar}
\usage{
split_gedcom_values(lines, char_limit)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}

\item{char_limit}{Maximum string length of values.}
}
\value{
A new character vector of gedcom lines, possibly expanded to include CONC/CONT lines.
}
\description{
This function introduces CONC/CONT lines for line values that exceed the given number of characters and
for lines containing line breaks.
}
