% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ordinance.R
\name{SpouseSealing}
\alias{SpouseSealing}
\title{Create a spouse sealing object}
\usage{
SpouseSealing(
  date = character(0),
  temple_name = character(0),
  place = NULL,
  ord_state = character(0),
  state_date = character(0),
  state_time = character(0),
  note_xrefs = character(0),
  notes = list(),
  citations = list()
)
}
\arguments{
\item{date}{The date given either as a formatted GEDCOM string, or a
\code{DateValue()} object.}

\item{temple_name}{The name of a temple of The Church of Jesus Christ of Latter-day Saints.}

\item{place}{The associated place. This can either be a
\code{Place()} object or a character string (a comma-separated string of region names,
ordered from smallest to largest).}

\item{ord_state}{An optional value from \verb{val_ordinance_states(@ord_type)}.}

\item{state_date}{The ordinance date given either as a formatted GEDCOM string, or a
\code{DateExact} object.}

\item{state_time}{The ordinance time given either as a formatted GEDCOM string, or a
\code{Time} object.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records.}
}
\value{
An S7 object representing a GEDCOM LDS_SPOUSE_SEALING.
}
\description{
Create a spouse sealing object
}
