% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_translation.R
\name{TranslationText}
\alias{TranslationText}
\title{Create a text translation object}
\usage{
TranslationText(
  text = character(0),
  language = character(0),
  media_type = character(0)
)
}
\arguments{
\item{text}{A character string. New lines are created with \\n.}

\item{language}{A character string of language tags as defined in BCP 47.}

\item{media_type}{The media type as defined in RFC 2045.}
}
\value{
An S7 object representing a GEDCOM text translation substructure.
}
\description{
Create a text translation object
}
