% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_siblings}
\alias{add_siblings}
\title{Create siblings for an Individual}
\usage{
add_siblings(x, xref, sexes = NULL)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{sexes}{A character string giving the sexes of each sibling. For example,
"FFM" to add two sisters and one brother.}
}
\value{
A gedcom object with additional sibling records.
}
\description{
Create siblings for an Individual
}
\details{
This function may also create a Family Group record and will
not modify existing siblings.
}
