% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_rm.R
\name{add_spouse}
\alias{add_spouse}
\title{Add a spouse record for an individual}
\usage{
add_spouse(x, xref, sex = "U", spou_name = NULL)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{sex}{The sex of the spouse.}

\item{spou_name}{Optional name to give to the spouse.
Surnames must be enclosed in forward slashes.}
}
\value{
A gedcom object with additional spouse and Family Group records.
}
\description{
This creates a record for a spouse and their Family record.
}
