% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_validators.R
\name{chk_input_choice}
\alias{chk_input_choice}
\title{Validate an input against a set of valid values.}
\usage{
chk_input_choice(input, choices)
}
\arguments{
\item{input}{The input.}

\item{choices}{A vector of valid values.}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate an input against a set of valid values.
}
\keyword{internal}
