% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_validators.R
\name{chk_input_S7classes}
\alias{chk_input_S7classes}
\title{Validate a list to ensure it contains elements of the correct S7 class}
\usage{
chk_input_S7classes(input, name, target_class, backup_pattern = NULL)
}
\arguments{
\item{input}{The input.}

\item{name}{The name of the input used in any error messages.}

\item{target_class}{The S7 class that the elements of the list should contain.}

\item{backup_pattern}{Need to handle List of mixed stuff vs character vector vs list of characters}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate a list to ensure it contains elements of the correct S7 class
}
