% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_validators.R
\name{chk_whole_number}
\alias{chk_whole_number}
\title{Validate an numeric input to be a whole number}
\usage{
chk_whole_number(input)
}
\arguments{
\item{input}{The input.}
}
\value{
Either a character string giving an error message, or NULL.
}
\description{
Validate an numeric input to be a whole number
}
\keyword{internal}
