% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_locations.R
\name{class_address}
\alias{class_address}
\title{Create an Address object}
\arguments{
\item{local_address_lines}{A character vector containing up to three local address lines.}

\item{city}{The city of the address.}

\item{state}{The state/county of the address.}

\item{postal_code}{The postal code of the address.}

\item{country}{The country of the address.}

\item{phone_numbers}{A character vector containing up to three phone numbers.}

\item{emails}{A character vector containing up to three email addresses.}

\item{faxes}{A character vector containing up to three fax numbers.}

\item{web_pages}{A character vector containing up to three web pages.}
}
\value{
An S7 Address object.
}
\description{
Create an Address object
}
