% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_association.R
\name{class_association}
\alias{class_association}
\title{Create an association object}
\usage{
class_association(
  indi_xref = class_missing,
  indi_phrase = class_missing,
  relation_is = class_missing,
  relation_phrase = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing
)
}
\arguments{
\item{indi_xref}{The cross-reference identifier of an individual record. If the individual
does not have a record, then this can be left blank and the value "@VOID@" will be used. However,
you will need to define an @indi_phrase.}

\item{indi_phrase}{Textual information that cannot be expressed in the @indi_xref.}

\item{relation_is}{The nature of the association. This must be a value from \code{val_roles()}.
If a value of "OTHER" is used, a @relation_phrase must be given.}

\item{relation_phrase}{Textual information that cannot be expressed in the relation.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}
}
\value{
An S7 object representing a GEDCOM ASSOCIATION_STRUCTURE.
}
\description{
Create an association object
}
