% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_fact_fam.R
\name{class_event_fam}
\alias{class_event_fam}
\title{Create a family event object}
\usage{
class_event_fam(
  fact_type = class_missing,
  fact_val = class_missing,
  fact_desc = class_missing,
  date = class_missing,
  place = class_missing,
  address = class_missing,
  phone_numbers = class_missing,
  emails = class_missing,
  faxes = class_missing,
  web_pages = class_missing,
  agency = class_missing,
  relig_affil = class_missing,
  cause = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  date_sort = class_missing,
  associations = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  media_links = class_missing,
  unique_ids = class_missing,
  husb_age = class_missing,
  husb_age_phrase = class_missing,
  wife_age = class_missing,
  wife_age_phrase = class_missing
)
}
\arguments{
\item{fact_type}{}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_value} object.}

\item{place}{see territory}

\item{phone_numbers}{}

\item{date_sort}{The date given either as a formatted GEDCOM string, or a
\code{class_date_sort} object.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Associated multimedia. This can either be a \code{class_media_link} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{unique_ids}{ds}
}
\value{
An S7 object representing a GEDCOM FAMILY_EVENT_STRUCTURE.
}
\description{
Create a family event object
}
