% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{class_facts_recorded}
\alias{class_facts_recorded}
\title{Create an object recording facts covered in a source record}
\usage{
class_facts_recorded(
  fact_types = class_missing,
  date_period = class_missing,
  date_phrase = class_missing,
  territory = class_missing
)
}
\arguments{
\item{fact_types}{A character string indicating the types
of events that were recorded in a particular source. Each event type is separated by a
comma and space. For example, a parish register of births, deaths, and marriages
would be BIRT, DEAT, MARR.}

\item{date_period}{A date period given either as a formatted GEDCOM string, or a
\code{class_date_period} object.}

\item{date_phrase}{Textual information that cannot be expressed in the date.}

\item{territory}{TODO}
}
\value{
An S7 object representing a GEDCOM SOUR.EVEN structure.
}
\description{
Create an object recording facts covered in a source record
}
