% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{class_gedcom_source}
\alias{class_gedcom_source}
\title{Create a GEDCOM source object}
\usage{
class_gedcom_source(
  product_id = class_missing,
  product_name = class_missing,
  product_version = class_missing,
  business_name = class_missing,
  business_address = class_missing,
  phone_numbers = class_missing,
  emails = class_missing,
  faxes = class_missing,
  web_pages = class_missing,
  data_name = class_missing,
  data_pubdate = class_missing,
  data_pubtime = class_missing,
  data_copyright = class_missing
)
}
\arguments{
\item{phone_numbers}{}

\item{data_copyright}{}
}
\value{
An S7 object representing a GEDCOM HEAD.SOUR.
}
\description{
Create a GEDCOM source object
}
