% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{class_gedcomS7}
\alias{class_gedcomS7}
\title{Create a GEDCOM object}
\usage{
class_gedcomS7(
  gedcom_version = class_missing,
  ext_tags = class_missing,
  source = class_missing,
  destination = class_missing,
  creation_date = class_missing,
  creation_time = class_missing,
  subm_xref = class_missing,
  gedcom_copyright = class_missing,
  default_language = class_missing,
  default_place_form = class_missing,
  notes = class_missing,
  note_xrefs = class_missing,
  update_change_dates = class_missing,
  add_creation_dates = class_missing,
  subm = class_missing,
  indi = class_missing,
  fam = class_missing,
  sour = class_missing,
  repo = class_missing,
  media = class_missing,
  note = class_missing,
  xref_prefixes = class_missing
)
}
\arguments{
\item{creation_date}{The creation date of the file given either as a formatted GEDCOM string, or a
\code{class_date_exact} object.}

\item{creation_time}{The creation time of the file given either as a formatted GEDCOM string, or a
\code{class_time} object.}

\item{subm_xref}{The cross-reference identifier of a submitter record.}

\item{default_place_form}{}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{update_change_dates}{Whether to automatically update change dates when updating records.
This happens when the record is pushed to the gedcom object.}

\item{add_creation_dates}{Whether to automatically add creation dates when creating records.
This happens when the record is pushed to the gedcom object.}
}
\value{
An S7 object representing a GEDCOM file.
}
\description{
Create a GEDCOM object
}
