% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_gedcom.R
\name{class_gedcomS7}
\alias{class_gedcomS7}
\title{Create a GEDCOM object}
\usage{
class_gedcomS7(
  gedcom_version = character(0),
  ext_tags = character(0),
  source = NULL,
  destination = character(0),
  creation_date = <object>,
  creation_time = character(0),
  subm_xref = character(0),
  gedcom_copyright = character(0),
  default_language = character(0),
  default_place_form = character(0),
  notes = list(),
  note_xrefs = character(0),
  update_change_dates = FALSE,
  add_creation_dates = FALSE,
  subm = list(),
  indi = list(),
  fam = list(),
  sour = list(),
  repo = list(),
  media = list(),
  note = list(),
  xref_prefixes = c(subm = "U", indi = "I", fam = "F", sour = "S", repo = "R", media =
    "M", note = "N")
)
}
\arguments{
\item{gedcom_version}{TODO}

\item{ext_tags}{TODO}

\item{source}{TODO}

\item{destination}{TODO}

\item{creation_date}{The creation date of the file given either as a formatted GEDCOM string, or a
\code{class_date_exact} object.}

\item{creation_time}{The creation time of the file given either as a formatted GEDCOM string, or a
\code{class_time} object.}

\item{subm_xref}{The cross-reference identifier of a submitter record.}

\item{gedcom_copyright}{TODO}

\item{default_language}{TODO}

\item{default_place_form}{TODO}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{update_change_dates}{Whether to automatically update change dates when updating records.
This happens when the record is pushed to the gedcom object.}

\item{add_creation_dates}{Whether to automatically add creation dates when creating records.
This happens when the record is pushed to the gedcom object.}

\item{subm, indi, fam, sour, repo, media, note}{A named list containing character vector representations
of GEDCOM records. Do not edit these parameters directly.}

\item{xref_prefixes}{A named vector containing any alphanumeric string (up to 6 characters long)
which will precede the number given to identify new records (of which there are 7 types).
This vector must be of a particular length with specific names. Default value:
c(subm = "U", indi = "I", fam = "F", sour = "S", repo = "R", media = "M", note = "N")
The order that these records appear in the vector will also dictate the order in which records
will appear in the exported file.}
}
\value{
An S7 object representing a GEDCOM file.
}
\description{
Create a GEDCOM object
}
