% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_personal_name.R
\name{class_name_pieces}
\alias{class_name_pieces}
\title{Create a name pieces object}
\usage{
class_name_pieces(
  prefix = class_missing,
  given = class_missing,
  nickname = class_missing,
  surname_prefix = class_missing,
  surname = class_missing,
  suffix = class_missing
)
}
\value{
An S7 object representing a GEDCOM PERSONAL_NAME_PIECES.
}
\description{
Create a name pieces object
}
