% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_note.R
\name{class_note}
\alias{class_note}
\title{Create a note structure object}
\usage{
class_note(
  text = class_missing,
  language = class_missing,
  media_type = class_missing,
  translations = class_missing,
  citations = class_missing
)
}
\arguments{
\item{text}{A character string. New lines are created with \\n.}

\item{language}{A character string of language tags as defined in BCP 47.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{translations}{A \code{class_translation_txt} object or a list of them.
One for each alternate translation of the text.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}
}
\value{
An S7 object representing a GEDCOM NOTE_STRUCTURE.
}
\description{
Create a note structure object
}
\details{
The shared note (SNOTE) alternative of this structure is defined
separately in relevant structures.
}
