% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_ordinance.R
\name{class_ordinance}
\alias{class_ordinance}
\title{Create an individual ordinance object}
\usage{
class_ordinance(
  ord_type = class_missing,
  date = class_missing,
  temple_name = class_missing,
  place = class_missing,
  ord_state = class_missing,
  state_date = class_missing,
  state_time = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  fam_xref = class_missing
)
}
\arguments{
\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_value} object.}

\item{place}{see territory}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{fam_xref}{The cross-reference identifier of a family record.}
}
\value{
An S7 object representing a GEDCOM LDS_INDIVIDUAL_ORDINANCE.
}
\description{
Create an individual ordinance object
}
