% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_locations.R
\name{class_place}
\alias{class_place}
\title{Create a Place object}
\arguments{
\item{name}{The jurisdictional name of the place where the event took place. Jurisdictions are separated
by a comma and space combination. For example: "Cove, Cache, Utah, United States of
America".
No part of the place name may be replaced by an abbreviation. Place names are not
terminated by a full stop or anything else.}

\item{phon_names}{A named vector of phonetic variations of the place name written in the same form as
\code{name}. The names of the vector contain the phonetisation method used for creating the phonetic text.}

\item{rom_names}{A named vector of romanised variations of the place name written in the same form as
\code{name}. The names of the vector contain the romanisation method used for creating the romanised text.}

\item{lat_long}{A character string giving the latitude and longitude separated by a space. See Details.}

\item{notes}{A list of \code{class_note} objects to record or reference notes about the place.}
}
\value{
An R7 Place object.
}
\description{
Create a Place object
}
\details{
The latitude coordinate is
the direction North or South from the equator in degrees and fraction of degrees carried out
to give the desired accuracy. For example: 18 degrees, 9 minutes, and 3.4 seconds North
would be formatted as N18.150944.

The longitude
coordinate is Degrees and fraction of degrees east or west of the zero or base meridian
coordinate. For example: 168 degrees, 9 minutes, and 3.4 seconds East would be formatted
as E168.150944.
}
