% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_indi.R
\name{class_record_indi}
\alias{class_record_indi}
\title{Create an individual record object}
\usage{
class_record_indi(
  xref = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  user_ids = class_missing,
  unique_ids = class_missing,
  ext_ids = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  media_links = class_missing,
  created = class_missing,
  updated = class_missing,
  pers_names = class_missing,
  sex = class_missing,
  facts = class_missing,
  non_events = class_missing,
  fam_links_chil = class_missing,
  fam_links_spou = class_missing,
  subm_xrefs = class_missing,
  associations = class_missing,
  alia_xrefs = class_missing,
  anci_xrefs = class_missing,
  desi_xrefs = class_missing
)
}
\arguments{
\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{user_ids}{fwe}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{dsfse}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{media_links}{Associated multimedia. This can either be a \code{class_media_link} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{updated}{}

\item{subm_xrefs}{A character vector of relevant submitter record cross-reference identifiers.}

\item{alia_xrefs}{A named character vector of relevant individual record cross-reference identifiers
who records also represent this individual. The vector names may provide a description of these records.}

\item{anci_xrefs}{A character vector of relevant submitter record cross-reference identifiers
who are interested in the ancestors of this individual.}

\item{desi_xrefs}{A character vector of relevant submitter record cross-reference identifiers
who are interested in the descendants of this individual.}
}
\value{
An S7 object representing a GEDCOM INDIVIDUAL_RECORD.
}
\description{
Create an individual record object
}
