% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_media.R
\name{class_record_media}
\alias{class_record_media}
\title{Create a multimedia record object}
\usage{
class_record_media(
  xref = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  user_ids = class_missing,
  unique_ids = class_missing,
  ext_ids = class_missing,
  created = class_missing,
  updated = class_missing,
  files = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing
)
}
\arguments{
\item{xref}{}

\item{ext_ids}{}

\item{files}{}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}
}
\value{
An S7 object representing a GEDCOM MULTIMEDIA_RECORD.
}
\description{
Create a multimedia record object
}
