% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{class_record_sour}
\alias{class_record_sour}
\title{Create a source record object}
\usage{
class_record_sour(
  xref = class_missing,
  confidential = class_missing,
  locked = class_missing,
  private = class_missing,
  user_ids = class_missing,
  unique_ids = class_missing,
  ext_ids = class_missing,
  note_xrefs = class_missing,
  notes = class_missing,
  citations = class_missing,
  media_links = class_missing,
  created = class_missing,
  updated = class_missing,
  facts_recorded = class_missing,
  agency = class_missing,
  data_note_xrefs = class_missing,
  data_notes = class_missing,
  originator = class_missing,
  full_title = class_missing,
  short_title = class_missing,
  publication_facts = class_missing,
  source_text = class_missing,
  repo_citations = class_missing
)
}
\arguments{
\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{user_ids}{A character vector of user-generated identifiers. The type of the identifiers can
be given in the vector names, e.g. c("Driving license number" = "ABC123")}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{ext_ids}{A named character vector of identifiers maintained by an external authority.
The names must be given as a URI. See the GEDCOM specification for more information.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Not used.}

\item{media_links}{Associated multimedia. This can either be a \code{class_media_link} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{created}{A \code{class_creation_date} object containing the date the record was created.
Creating an object with no parameters sets the date to today.}

\item{updated}{A \code{class_change_date} object containing the date the record was updated.
Creating an object with no parameters sets the date to today.}

\item{facts_recorded}{The facts recorded by the source. This can either be a \code{class_facts_recorded} object,
a list of them, or a character vector of comma-delimited fact types. If a character vector is
provided then only the fact types themselves can be recorded (and not associated information).
This option is easier if associated information is not required. For example, a parish register of
births, deaths, and marriages would be "BIRT, DEAT, MARR". The \code{val_fact_types()} function gives a
list of possible fact types.}

\item{originator}{The person, agency, or entity who created the record. For a published work,
this could be the author, compiler, transcriber, abstractor, or editor. For an unpublished
source, this may be an individual, a government agency, church organization, or private organization.}

\item{full_title}{The full title of the source.}

\item{short_title}{A shortened name of the source used for sorting, filing, and retrieving records.}

\item{publication_facts}{When and where the record was created. For published works, this
includes information such as the city of publication, name of the publisher, and year of publication.}

\item{source_text}{A verbatim copy of any description contained within the source. This can
either be a \code{class_translation_txt} object, a list of them, or a character vector of text.
If any \code{class_translation_txt} objects are provided, then you will be forced to define an
associated language or media type.}

\item{repo_citations}{Associated repositories. This can either be a \code{class_repository_citation} object,
a list of them, or a character vector of XREFs of repository records. If a character vector is
provided then only the XREFs themselves can be recorded (and not associated information).
This option is easier if associated information is not required.}
}
\value{
An S7 object representing a GEDCOM SOURCE_RECORD.
}
\description{
Create a source record object
}
