% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{class_source_call_number}
\alias{class_source_call_number}
\title{Create a source call number object}
\usage{
class_source_call_number(
  call_number = class_missing,
  medium = class_missing,
  medium_phrase = class_missing
)
}
\value{
An S7 object representing the CALN substructure of a GEDCOM
SOURCE_REPOSITORY_CITATION.
}
\description{
Create a source call number object
}
