% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Identify all ancestors for an individual}
\usage{
get_ancestors(
  x,
  xref,
  inc_indi = FALSE,
  inc_sibs = FALSE,
  inc_famg = FALSE,
  inc_supp = FALSE,
  birth_only = TRUE
)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{inc_indi}{Whether to also include the individual themselves.}

\item{inc_sibs}{Whether to also include all siblings of ancestors (siblings of this individual will only be
included if the individual is included).}

\item{inc_famg}{Whether to also include all Family Group records where this individual is a child
(and all ancestors' Family Group records).}

\item{inc_supp}{Whether to also include all supporting records (Note, Source, Repository, Multimedia).}

\item{birth_only}{Whether to only include biological ancestors.}
}
\value{
A vector of xrefs of ancestors.
}
\description{
This function identifies records in an entire branch of the family tree above a certain individual.
}
