% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_parents}
\alias{get_indi_parents}
\title{Identify all parents for an individual}
\usage{
get_indi_parents(x, xref, birth_only = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{birth_only}{Whether to only return biological parents.}
}
\value{
A character vector of parent xrefs.
}
\description{
Identify all parents for an individual
}
