% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_partners}
\alias{get_indi_partners}
\title{Identify all partners for an individual}
\usage{
get_indi_partners(x, xref)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}
}
\value{
A character vector of xrefs.
}
\description{
Identify all partners for an individual
}
