% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_indi_siblings}
\alias{get_indi_siblings}
\title{Identify all siblings for an individual}
\usage{
get_indi_siblings(x, xref, birth_only = FALSE, inc_half = FALSE)
}
\arguments{
\item{x}{A gedcom object.}

\item{xref}{The xref of an Individual record.}

\item{birth_only}{Whether to only return biological siblings.}

\item{inc_half}{Whether to include siblings that only share one parent.}
}
\value{
A character vector of sibling xrefs.
}
\description{
Identify all siblings for an individual
}
