% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_media.R
\name{MediaFile}
\alias{MediaFile}
\title{Create a media file object}
\usage{
MediaFile(
  location = character(0),
  title = character(0),
  media_type = character(0),
  medium = character(0),
  medium_phrase = character(0),
  media_alt = character(0)
)
}
\arguments{
\item{location}{An absolute or relative URL to the file.}

\item{title}{The title of the multimedia record. This will supersede any title given
in the record.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{medium}{TODO}

\item{medium_phrase}{TODO}

\item{media_alt}{A named vector of the media in alternative media forms, c(form = location)}
}
\value{
An S7 object representing a GEDCOM multimedia file substructure.
}
\description{
Create a media file object
}
