% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_personal_name.R
\name{PersonalName}
\alias{PersonalName}
\title{Create a personal name object}
\usage{
PersonalName(
  pers_name = character(0),
  name_type = character(0),
  type_phrase = character(0),
  name_pieces = NULL,
  name_translations = list(),
  notes = list(),
  note_xrefs = character(0),
  citations = list()
)
}
\arguments{
\item{pers_name}{The full name of the individual. Surnames should be enclosed in forward slashes.}

\item{name_type}{An optional name type taken from \code{val_name_types()}.}

\item{type_phrase}{An optional free text description of the name type. This
is required if the name type is "OTHER".}

\item{name_pieces}{A \code{PersonalNamePieces()} object defining the pieces of the full name.}

\item{name_translations}{A \code{PersonalNameTran()} object or a list of them, providing
different translations of this personal name.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records.}
}
\value{
An S7 object representing a GEDCOM PERSONAL_NAME_STRUCTURE.
}
\description{
Create a personal name object
}
