% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{prepare_gedcom_lines}
\alias{prepare_gedcom_lines}
\title{Prepare GEDCOM lines for export}
\usage{
prepare_gedcom_lines(lines)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}
}
\value{
A vector of GEDCOM lines ready for export.
}
\description{
Prepare GEDCOM lines for export
}
