% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_class_props.R
\name{prop_definitions}
\alias{prop_definitions}
\title{Define properties in GEDCOM 7.0 specification}
\arguments{
\item{text}{A character string. New lines are created with \\n.}

\item{language}{A character string of language tags as defined in BCP 47.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{translations}{A \code{TranslationText()} object or a list of them.
One for each alternate translation of the text.}

\item{title}{The title of the multimedia record.}

\item{source_text}{A verbatim copy of any description contained within the source. This can
either be a \code{TranslationText()} object, a list of them, or a character vector of text.
If any \code{TranslationText()} objects are provided, then you will be forced to define an
associated language or media type.}

\item{fact_type}{TODO}

\item{medium}{TODO}

\item{medium_phrase}{TODO}

\item{media_alt}{A named vector of the media in alternative media forms, c(form = location)}

\item{pers_name}{TODO}

\item{name_pieces}{A \code{PersonalNamePieces()} object defining the pieces of the full name.}

\item{event_type}{TODO}

\item{fact_val}{TODO}

\item{fact_desc}{TODO}

\item{place}{see territory}

\item{address}{The address given either as a \code{Address()} object or as a character string.
This would be as written on a mailing label with new lines separated by \\n.}

\item{phone_numbers}{A character vector of phone numbers.}

\item{emails}{A character vector of email addresses.}

\item{faxes}{A character vector of fax numbers.}

\item{web_pages}{A character vector of web page URLs.}

\item{agency}{The organization, institution, corporation, person, or other entity that
has responsibility for the associated fact. Examples are an employer of a person of an
associated occupation, or an educational establishment that has awarded a scholastic award.}

\item{relig_affil}{A religious denomination associated with the fact.}

\item{cause}{Used in special cases to record the reasons which precipitated the fact (e.g. cause of death).}

\item{confidential}{A logical value indicating whether the associated record/fact should be
treated as confidential. This allows them to be excluded on export.}

\item{locked}{A logical value indicating whether the associated record/fact should be
treated as read-only.}

\item{private}{A logical value indicating whether the associated record/fact should be
treated as private. This allows them to be excluded on export.}

\item{associations}{Associated individuals. This can either be a \code{Association()} object or a list of them.}

\item{age}{A character string that indicates the age in years, months, weeks and/or days
that the individual was at the time of the fact. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 1w 3d".
Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age
in the corresponding phrase parameter.}

\item{husb_age}{A character string that indicates the age in years, months, weeks and/or days
that the husband was at the time of the fact. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 1w 3d".
Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age
in the corresponding phrase parameter.}

\item{wife_age}{A character string that indicates the age in years, months, weeks and/or days
that the wife was at the time of the fact. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 1w 3d".
Age bounds can also be included, for example; "< 40y". If the age doesn't fit this format then describe the age
in the corresponding phrase parameter.}

\item{age_phrase}{Free text information that cannot be expressed in the individual's age.}

\item{husb_age_phrase}{Free text information that cannot be expressed in the husband's age.}

\item{wife_age_phrase}{Free text information that cannot be expressed in the wife's age.}

\item{temple_name}{The name of a temple of The Church of Jesus Christ of Latter-day Saints.}

\item{ord_state}{An optional value from \verb{val_ordinance_states(@ord_type)}.}

\item{state_date}{The ordinance date given either as a formatted GEDCOM string, or a
\code{DateExact} object.}

\item{state_time}{The ordinance time given either as a formatted GEDCOM string, or a
\code{Time} object.}

\item{xref}{The cross-reference identifier for this record. You should not edit this at all
as maintenance of these is done automatically.}

\item{fam_xref}{The cross-reference identifier of a family record.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{subm_xrefs}{A character vector of relevant submitter record cross-reference identifiers.}

\item{unique_ids}{A character vector of enduring and globally-unique identifiers. These need
to be formatted in line with RFC 4122 and can be generated with \code{uuid::UUIDgenerate()}.}

\item{user_ids}{A character vector of user-generated identifiers. The type of the identifiers can
be given in the vector names, e.g. c("Driving license number" = "ABC123")}

\item{ext_ids}{A named character vector of identifiers maintained by an external authority.
The names must be given as a URI. See the GEDCOM specification for more information.}

\item{media_links}{Associated multimedia. This can either be a \code{MediaLink()} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{SourceCitation()} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{year}{The year given as an integer (greater than 0).}

\item{month}{The month of the year given as an integer between 1 and 12.}

\item{day}{The day of the month given as an integer between 1 and 31.}

\item{date_exact}{An exact date given either as a formatted GEDCOM string, or a
\code{DateExact()} object. If not given, it will default to today's date.}

\item{date_period}{A date period given either as a formatted GEDCOM string, or a
\code{DatePeriod()} object.}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{DateValue()} object.}

\item{date_sort}{The date given either as a formatted GEDCOM string, or a
\code{DateSorting()} object.}

\item{date_phrase}{Textual information that cannot be expressed in the date.}

\item{start_date}{The start of the period/range given either as a formatted GEDCOM string, or a
\code{DateGregorian()} object.}

\item{end_date}{The end of the period/range given either as a formatted GEDCOM string, or a
\code{DateGregorian()} object.}

\item{time}{The time given either as a formatted GEDCOM string, or a
\code{Time()} object.}

\item{created}{A \code{CreationDate()} object containing the date the record was created.
Creating an object with no parameters sets the date to today.}

\item{updated}{A \code{ChangeDate()} object containing the date the record was updated.
Creating an object with no parameters sets the date to today.}
}
\description{
Define common properties used in S7 classes.
}
\details{
This empty function serves as a single location where all shared properties are defined,
mainly for efficiency and maintainability.
}
\keyword{internal}
