% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_class_props.R
\name{prop_definitions}
\alias{prop_definitions}
\title{Define properties in GEDCOM 7.0 specification}
\arguments{
\item{full}{A full address as it would appear on a mailing label, with lines separated
by \n. For example:
"The White House\n1600 Pennsylvania Avenue, N.W.\nWashington, DC 20500\nUnited States of America"}

\item{city}{The city component of the address.}

\item{state}{The state component of the address.}

\item{postal_code}{The postal code component of the address.}

\item{country}{The country component of the address.}

\item{text}{A character string. New lines are created with \n.}

\item{language}{Language tags as defined in BCP 47.}

\item{media_type}{The media type as defined in RFC 2045.}

\item{translations}{A \code{class_translation_txt} object or a list of them.
One for each alternate translation of the text.}

\item{title}{The title of the multimedia record. This will supersede any title given
in the record.}

\item{crop}{Whether to crop the multimedia to a specific area.}

\item{top}{The number of pixels to omit from the top side of the image.}

\item{left}{The number of pixels to omit from the left side of the image.}

\item{height}{The height in pixels of the cropped region.}

\item{width}{The width in pixels of the cropped region.}

\item{where}{A specific location within the information referenced. For a published work, this could
include the volume of a multi-volume work and the page number or numbers. For a
periodical, it could include volume, issue, and page numbers. For a newspaper, it could
include a date, page number, and column number. For an unpublished source or micro‐
filmed works, this could be a film or sheet number, page number, or frame number. A
census record might have an enumerating district, page number, line number, dwelling
number, and family number.
It is recommended that the data in this field be formatted comma-separated with label:
value pairs}

\item{source_text}{A verbatim copy of any description contained within the source. This can
either be a \code{class_translation_txt} object, a list of them, or a character vector of text.
If any \code{class_translation_txt} objects are provided, then you will be forced to define an
associated language or media type.}

\item{fact_type}{}

\item{fact_phrase}{Textual information that cannot be expressed in the fact type.}

\item{role}{What role this person played in this fact.}

\item{role_phrase}{Textual information that cannot be expressed in the role.}

\item{certainty}{An enumerated value indicating the credibility of a
piece of information, based on its supporting evidence. Some systems use this feature to
rank multiple conflicting opinions for display of most likely information first. It is not
intended to eliminate the receivers’ need to evaluate the evidence for themselves.
"0" = unreliable/estimated data
"1" = Questionable reliability of evidence
"2" = Secondary evidence, data officially recorded sometime after event
"3" = Direct and primary evidence used, or by dominance of the evidence}

\item{indi_phrase}{Textual information that cannot be expressed in the @indi_xref.}

\item{fact_types}{A character string indicating the types
of events that were recorded in a particular source. Each event type is separated by a
comma and space. For example, a parish register of births, deaths, and marriages
would be BIRT, DEAT, MARR.}

\item{media_alt}{A named vector of the media in alternative media forms, c(form = location)}

\item{place}{see territory}

\item{phone_numbers}{}

\item{default_place_form}{}

\item{xref}{}

\item{indi_xref}{The cross-reference identifier of an individual record. If the individual
does not have a record, then this can be left blank and the value "@VOID@" will be used. However,
you will need to define an @indi_phrase.}

\item{fam_xref}{The cross-reference identifier of a family record.}

\item{sour_xref}{The cross-reference identifier of a source record. If the source
does not have a record, then this can be left blank and the value "@VOID@" will be used. However,
you will need to describe the source in @where.}

\item{repo_xref}{The cross-reference identifier of a repository record. If the repository
does not have a record, then this can be left blank and the value "@VOID@" will be used. However,
you will need to describe the repository in @notes.}

\item{media_xref}{The cross-reference identifier of a multimedia record.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{husb_xref}{named vector}

\item{wife_xref}{named vector}

\item{chil_xrefs}{named vector}

\item{desi_xrefs}{}

\item{unique_ids}{ds}

\item{user_ids}{fwe}

\item{ext_ids}{dsfse}

\item{media_links}{Associated multimedia. This can either be a \code{class_media_link} object, a list of them,
or a character vector of XREFs of multimedia records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{notes}{Associated notes. This can either be a \code{class_note} object, a list of them,
or a character vector of notes. If a character vector is provided then only the notes themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{citations}{Associated sources. This can either be a \code{class_citation} object, a list of them,
or a character vector of XREFs of source records. If a character vector is provided then only the XREFs themselves
can be recorded (and not associated information). This option is easier if
associated information is not required.}

\item{hour}{The hour of the day given as an integer between 0 and 23.}

\item{minute}{The minute of the hour given as an integer between 0 and 59.}

\item{second}{The second of the minute given as an integer between 0 and 59.}

\item{fraction}{The fraction of the second given as
an integer.}

\item{utc}{Whether the time is in Coordinated Universal Time (UTC) (TRUE, the default) or
is in local time (FALSE).}

\item{year}{The year given as an integer (greater than 0).}

\item{month}{The month of the year given as an integer between 1 and 12.}

\item{day}{The day of the month given as an integer between 1 and 31.}

\item{bce}{Whether the date is Before the Common Era. This is FALSE by default,
but if TRUE, only the year should be given.}

\item{about}{Whether the date is near to the date given.}

\item{calc}{Whether the date is calculated from other values.}

\item{est}{Whether the date is near to the date given, and is calculated from other values.}

\item{date_greg}{A Gregorian date given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{date_exact}{An exact date given either as a formatted GEDCOM string, or a
\code{class_date_exact} object. If not given, it will default to today's date.}

\item{date_period}{A date period given either as a formatted GEDCOM string, or a
\code{class_date_period} object.}

\item{date}{The date given either as a formatted GEDCOM string, or a
\code{class_date_value} object.}

\item{date_sort}{The date given either as a formatted GEDCOM string, or a
\code{class_date_sort} object.}

\item{date_phrase}{Textual information that cannot be expressed in the date.}

\item{start_date}{The start of the period/range given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{end_date}{The end of the period/range given either as a formatted GEDCOM string, or a
\code{class_date_greg} object.}

\item{time}{The time given either as a formatted GEDCOM string, or a
\code{class_time} object.}

\item{updated}{}

\item{add_creation_dates}{}
}
\description{
Define all properties used in S7 classes.
}
\details{
This empty function serves as a single location where all properties are defined,
mainly for efficiency and maintainability.
}
