% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_import.R
\name{read_gedcom}
\alias{read_gedcom}
\title{Import a GEDCOM file}
\usage{
read_gedcom(filepath = file.choose())
}
\arguments{
\item{filepath}{The full filepath of the GEDCOM file.}
}
\value{
A gedcom S7 object.
}
\description{
Imports a *.ged file and creates a gedcom object.
}
