% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_record_sour.R
\name{RepositoryCitation}
\alias{RepositoryCitation}
\title{Create a repository citation object}
\usage{
RepositoryCitation(
  repo_xref = "@VOID@",
  notes = list(),
  note_xrefs = character(0),
  call_numbers = list()
)
}
\arguments{
\item{repo_xref}{The cross-reference identifier of a repository record. If the repository
does not have a record, then this can be left blank and a void xref will be used. However,
you should describe the repository in @notes.}

\item{notes}{Associated notes. This can either be a \code{Note()} object, a list of them,
or a character vector of notes.}

\item{note_xrefs}{A character vector of relevant note record cross-reference identifiers.}

\item{call_numbers}{Call number(s) used to file and retrieve items from the repository.
This can either be a \code{SourceCallNumber} object, a list of them,
or a character vector of call numbers.}
}
\value{
An S7 object representing a GEDCOM SOURCE_REPOSITORY_CITATION.
}
\description{
Create a repository citation object
}
