% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_rm.R
\name{rm_records}
\alias{rm_records}
\title{Remove records from a GEDCOM object}
\usage{
rm_records(x, xrefs)
}
\arguments{
\item{x}{A gedcom object.}

\item{xrefs}{A character vector of xrefs to remove.}
}
\value{
The gedcom object with the records removed. Pointers to the record will
be replaced with @VOID@.
}
\description{
Remove records from a GEDCOM object
}
