% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{split_gedcom_values}
\alias{split_gedcom_values}
\title{Create continuation lines}
\usage{
split_gedcom_values(lines)
}
\arguments{
\item{lines}{A character vector of gedcom lines.}
}
\value{
A new character vector of gedcom lines, possibly expanded to include CONT lines.
}
\description{
This function introduces CONT lines for line values that contain line breaks.
}
