% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lookups.R
\name{val_record_types}
\alias{val_record_types}
\alias{val_adoptive_parents}
\alias{val_individual_attribute_types}
\alias{val_individual_event_types}
\alias{val_fact_types}
\alias{val_individual_ordinance_types}
\alias{val_family_ordinance_types}
\alias{val_ordinance_states}
\alias{val_medium_types}
\alias{val_pedigree_types}
\alias{val_certainty}
\alias{val_restriction}
\alias{val_roles}
\alias{val_sexes}
\alias{val_confidence_types}
\alias{val_name_types}
\title{Lookup values}
\usage{
val_record_types()

val_adoptive_parents()

val_individual_attribute_types(inc_generic = FALSE)

val_individual_event_types(inc_generic = FALSE)

val_fact_types(inc_generic = FALSE)

val_individual_ordinance_types()

val_family_ordinance_types()

val_ordinance_states(ord_type)

val_medium_types()

val_pedigree_types()

val_certainty()

val_restriction()

val_roles()

val_sexes()

val_confidence_types()

val_name_types()
}
\arguments{
\item{inc_generic}{Whether to include a generic fact.}

\item{ord_type}{One of the values in \code{val_individual_ordinance_types()} or
\code{val_family_ordinance_types()}.}
}
\value{
A vector of allowed values.
}
\description{
Lookup values
}
